API

Index

DataFrames.DataFrame

DataFrames.DataFrame

DataFrames.DataFrame

DataFrames.DataFrame

PlantSimEngine.AbstractModel

PlantSimEngine.ModelList

PlantSimEngine.Status

PlantSimEngine.TimeStepTable

Base.copy

Base.copy

Base.copy

Base.getindex

PlantSimEngine.EF

PlantSimEngine.NRMSE

PlantSimEngine.RMSE

PlantSimEngine.add_model_vars

PlantSimEngine.convert_vars

PlantSimEngine.dep

PlantSimEngine.diff_vars

PlantSimEngine.dr

PlantSimEngine.draw_dependency_tree

PlantSimEngine.draw_guide

PlantSimEngine.draw_panel

PlantSimEngine.get_attr_i

PlantSimEngine.homogeneous_ts_kwargs

PlantSimEngine.homogeneous_ts_kwargs

PlantSimEngine.init_mtg_models!

PlantSimEngine.init_status!

PlantSimEngine.init_variables

PlantSimEngine.init_variables_manual

PlantSimEngine.inputs

PlantSimEngine.is_initialized

PlantSimEngine.outputs

PlantSimEngine.pre_allocate_attr!

PlantSimEngine.pull_status!

PlantSimEngine.pull_status_one_step!

PlantSimEngine.status

PlantSimEngine.to_initialize

PlantSimEngine.update_mtg_models!

PlantSimEngine.variables

PlantSimEngine.variables

PlantSimEngine.variables_typed

PlantSimEngine.vars_not_init_

PlantSimEngine.@gen_process_methods

API documentation

DataFrames.DataFrame — Method

DataFrame(components::T) where {T<:ModelList}

Generic implementation of DataFrame for a single ModelList model.

DataFrames.DataFrame — Method

DataFrame(components <: AbstractArray{<:ModelList})
DataFrame(components <: AbstractDict{N,<:ModelList})

Fetch the data from a ModelList (or an Array/Dict of) status into a DataFrame.

DataFrames.DataFrame — Method

DataFrame(components::ModelList{T,S}) where {T,S<:AbstractDict}

Implementation of DataFrame for a ModelList model with one time step.

DataFrames.DataFrame — Method

DataFrame(components::ModelList{T,<:TimeStepTable})

Implementation of DataFrame for a ModelList model with several time steps.

PlantSimEngine.AbstractModel — Type

Abstract model type. All models are subtypes of this one, see e.g. AbstractAModel

PlantSimEngine.ModelList — Type

ModelList(models::M, status::S)
ModelList(;
    status=nothing,
    init_fun::Function=init_fun_default,
    type_promotion=nothing,
    variables_check=true,
    kwargs...
)

List the models for a simulation (models), and does all boilerplate for variable initialization, type promotion, time steps handling.

Note

The status field depends on the input models. You can get the variables needed by a model using variables on the instantiation of a model. You can also use inputs and outputs instead.

Arguments

- `models`: a list of models. Usually given as a `NamedTuple`, but can be any other structure that 
implements `getproperty`.
- `status`: a structure containing the initializations for the variables of the models. Usually a NamedTuple
when given as a kwarg, or any structure that implements the Tables interface from `Tables.jl` (*e.g.* DataFrame).
- `init_fun`: a function that initializes the status based on a vector of NamedTuples (see details).
- `type_promotion`: optional type conversion for the variables with default values.
`nothing` by default, *i.e.* no conversion. Note that conversion is not applied to the
variables input by the user as `kwargs` (need to do it manually).
Should be provided as a Dict with current type as keys and new type as values.
- `variables_check=true`: check that all needed variables are initialized by the user.
- `kwargs`: the models, named after the process they simulate.

Details

The argument init_fun is set by default to init_fun_default which initializes the status with a TimeStepTable of Status structures.

If you change init_fun by another function, make sure the type you are using (i.e. in place of TimeStepTable) implements the Tables.jl interface (e.g. DataFrame does). And if you still use TimeStepTable but only change Status, make sure the type you give is indexable using the dot synthax (e.g. x.var).

If you need to input a custom Type for the status and make your users able to only partially initialize the status field in the input, you'll have to implement a method for add_model_vars!, a function that adds the models variables to the type in case it is not fully initialized.

Examples

A leaf with a width of 0.03 m, that uses the Monteith and Unsworth (2013) model for energy balance, the Farquhar et al. (1980) model for photosynthesis, and a constant stomatal conductance for CO₂ of 0.0011 with no residual conductance.

using PlantBiophysics

leaf = ModelList(
    energy_balance = Monteith(),
    photosynthesis = Fvcb(),
    stomatal_conductance = ConstantGs(0.0, 0.0011)
)

No variables were given as keyword arguments, that means that the status of the leaf is not set yet, and all variables are initialized to typemin(Type), i.e. -Inf for floating point numbers. This component cannot be simulated yet.

To know which variables we need to initialize for a simulation, we use to_initialize:

to_initialize(leaf)

The meaning and units of the variables can be found on the documentation of each model, e.g. here for photosynthesis.

We can now provide values for these variables:

leaf = ModelList(
    energy_balance = Monteith(),
    photosynthesis = Fvcb(),
    stomatal_conductance = ConstantGs(0.0, 0.0011),
    status = (Rₛ = 13.747, sky_fraction = 1.0, d = 0.03, PPFD = 1500)
)

We can now simulate the leaf, e.g. for the energy_balance (coupled to photosynthesis and stomatal conductance):

meteo = Atmosphere(T = 22.0, Wind = 0.8333, P = 101.325, Rh = 0.4490995)

energy_balance!(leaf,meteo)

DataFrame(leaf)

If we want to use special types for the variables, we can use the type_promotion argument:

leaf = ModelList(
    energy_balance = Monteith(),
    photosynthesis = Fvcb(),
    stomatal_conductance = ConstantGs(0.0, 0.0011),
    status = (Rₛ = 13.747, sky_fraction = 1.0, d = 0.03, PPFD = 1500),
    type_promotion = Dict(Float64 => Float32)
)

We can also use DataFrame as the status type:

using DataFrames
df = DataFrame(:Rₛ => [13.747, 13.8], :sky_fraction => [1.0, 1.0], :d => [0.03, 0.03], :PPFD => [1300.0, 1500.0])
m = ModelList(
    energy_balance=Monteith(),
    photosynthesis=Fvcb(),
    stomatal_conductance=Medlyn(0.03, 12.0),
    status=df,
    init_fun=x -> DataFrame(x)
)

# Note that we use `init_fun` to force the status into a `DataFrame`,
# otherwise it would be automatically converted into a `TimeStepTable{Status}`.

Note that computations will be slower using DataFrame, so if performance is an issue, use TimeStepTable instead (or a NamedTuple as shown in the example).

PlantSimEngine.Status — Type

Status(vars)

Status type used to store the values of the variables during simulation. It is mainly used as the structure to store the variables in the TimeStepRow of a TimeStepTable of a ModelList.

Most of the code is taken from MasonProtter/MutableNamedTuples.jl, so Status is a MutableNamedTuples with a few modifications, so in essence, it is a stuct that stores a NamedTuple of the references to the values of the variables, which makes it mutable.

Examples

# A leaf with one value for all variables will make a status with one time step:
st = Status(Rₛ=13.747, sky_fraction=1.0, d=0.03, PPFD=1500.0)

# Indexing a Status with a symbol returns the value of the variable:
st[:Tₗ]

# Indexing the Status with an integer returns the value of the variable by position:
st[1]

PlantSimEngine.TimeStepTable — Type

TimeStepTable(vars)
TimeStepTable{Status}(df::DataFrame)

TimeStepTable stores the values of the variables for each time step of a simulation. For example, it is used as the structure in the status field of the ModelList type.

TimeStepTable implements the Tables.jl interface, so it can be used with any package that uses Tables.jl (like DataFrames.jl).

Examples

# A leaf with several values for at least one of its variable will automatically use 
# TimeStepTable{Status} with the time steps:
leaf = ModelList(
    photosynthesis = Fvcb(),
    stomatal_conductance = Medlyn(0.03, 12.0),
    status=(Tₗ=[25.0, 26.0], PPFD=1000.0, Cₛ=400.0, Dₗ=1.0)
)

# The status of the leaf is a TimeStepTable:
status(leaf)

# Of course we can also create a TimeStepTable manually:
TimeStepTable(
    [
        Status(Tₗ=25.0, PPFD=1000.0, Cₛ=400.0, Dₗ=1.0),
        Status(Tₗ=26.0, PPFD=1200.0, Cₛ=400.0, Dₗ=1.2),
    ]
)

# We can also create a TimeStepTable from a DataFrame:
using DataFrames
df = DataFrame(
    Tₗ=[25.0, 26.0],
    PPFD=[1000.0, 1200.0],
    Cₛ=[400.0, 400.0],
    Dₗ=[1.0, 1.2],
)
TimeStepTable{Status}(df)

Base.copy — Method

Base.copy(l::AbstractArray{<:ModelList})

Copy an array-alike of ModelList

Base.copy — Method

Base.copy(l::AbstractDict{N,<:ModelList} where N)

Copy a Dict-alike ModelList

Base.copy — Method

Base.copy(l::ModelList)
Base.copy(l::ModelList, status)

Copy a ModelList, eventually with new values for the status.

Base.getindex — Method

getindex(component<:ModelList, key::Symbol)
getindex(component<:ModelList, key)

Indexing a component models structure: - with an integer, will return the status at the ith time-step - with anything else (Symbol, String) will return the required variable from the status

Examples

lm = ModelList(
    energy_balance = Monteith(),
    photosynthesis = Fvcb(),
    stomatal_conductance = ConstantGs(0.0, 0.0011),
    status = (Cᵢ = 380.0, Tₗ = [20.0, 25.0])
)

lm[:Tₗ] # Returns the value of the Tₗ variable
lm[2]  # Returns the status at the second time-step
lm[2][:Tₗ] # Returns the value of Tₗ at the second time-step
lm[:Tₗ][2] # Equivalent of the above

PlantSimEngine.EF — Method

EF(obs,sim)

Returns the Efficiency Factor between observations `obs` and simulations `sim` using NSE (Nash-Sutcliffe efficiency) model.
More information can be found at https://en.wikipedia.org/wiki/Nash%E2%80%93Sutcliffe_model_efficiency_coefficient.

The closer to 1 the better.

PlantSimEngine.NRMSE — Method

NRMSE(obs,sim)

Returns the Normalized Root Mean Squared Error between observations `obs` and simulations `sim`.
Normalization is performed using division by observations range (max-min).

Output: Float/Particles

PlantSimEngine.RMSE — Method

RMSE(obs,sim)

Returns the Root Mean Squared Error between observations `obs` and simulations `sim`.

The closer to 0 the better.

PlantSimEngine.add_model_vars — Method

add_model_vars(x)

Check which variables in x are not initialized considering a set of models and the variables needed for their simulation. If some variables are unitialized, initialize them to their default values.

This function needs to be implemented for each type of x (please do it if you need it).

Careful, the function mutates x in place for performance. We don't put the ! in the name just because it also returns it (impossible to mutate when x is nothing)

PlantSimEngine.convert_vars — Method

convert_vars(type_promotion::Dict{DataType,DataType}, ref_vars)
convert_vars(type_promotion::Nothing, ref_vars)

Convert the status variables to the type specified in the type promotion dictionary.

Examples

If we want all the variables that are Reals to be Float32, we can use:

ref_vars = init_variables(energy_balance=Monteith(), photosynthesis=Fvcb(), stomatal_conductance=Medlyn(0.03, 12.0))
type_promotion = Dict(Real => Float32)

convert_vars(type_promotion, ref_vars)

PlantSimEngine.dep — Method

dep(models::ModelList; verbose::Bool=true)

Get the model dependency tree given a ModelList. If one tree is returned, then all models are coupled. If several trees are returned, then only the models inside each tree are coupled, and the models in different trees are not coupled.

PlantSimEngine.diff_vars — Method

diff_vars(x, y)

Returns the names of variables that have different values in x and y.

PlantSimEngine.dr — Method

dr(obs,sim)

Returns the Willmott’s refined index of agreement dᵣ.
Willmot et al. 2011. A refined index of model performance. https://rmets.onlinelibrary.wiley.com/doi/10.1002/joc.2419

The closer to 1 the better.

PlantSimEngine.draw_dependency_tree — Method

draw_dependency_tree(
    tree, node;
    guides_style::String=TERM_THEME[].tree_guide_style,
    dep_tree_guides=(space=" ", vline="│", branch="├", leaf="└", hline="─")
)

Draw the dependency tree.

PlantSimEngine.draw_guide — Method

draw_guide(h, w, prefix, isleaf, guides)

Draw the line guide for one node of the dependency tree.

PlantSimEngine.draw_panel — Method

draw_panel(node, tree, prefix, dep_tree_guides)

Draw the panels for all dependencies

PlantSimEngine.get_attr_i — Method

get_attr_i(node, attr, i<:Nothing)
get_attr_i(node, attr, i)

Get node attribute at ith value or only value if i<:Nothing

PlantSimEngine.homogeneous_ts_kwargs — Method

homogeneous_ts_kwargs(kwargs)

By default, the function returns its argument.

PlantSimEngine.homogeneous_ts_kwargs — Method

kwargs_to_timestep(kwargs::NamedTuple{N,T}) where {N,T}

Takes a NamedTuple with optionnaly vector of values for each variable, and makes a vector of NamedTuple, with each being a time step. It is used to be able to e.g. give constant values for all time-steps for one variable.

Examples

homogeneous_ts_kwargs((Tₗ=[25.0, 26.0], PPFD=1000.0))
# [(Tₗ=25.0, PPFD=1000.0), (Tₗ=26.0, PPFD=1000.0)]

PlantSimEngine.init_mtg_models! — Function

init_mtg_models!(
    mtg::MultiScaleTreeGraph.Node,
    models::Dict{String,<:ModelList},
    i=nothing;
    verbose=true,
    attr_name=:models
)

initialize the components of an MTG (i.e. nodes) with the corresponding models.

The function checks if the models associated to each component of the MTG are fully initialized, and if not, tries to initialize the variables using the MTG attributes with the exact same name, and if not found, returns an error.

Arguments

mtg::MultiScaleTreeGraph.Node: the MTG tree.

models::Dict{String,ModelList}: a dictionary of models named by components names

i=nothing: the time-step to initialize. If nothing, initialize all the time-steps.

verbose = true: return information during the processes

attr_name = :models: the node attribute name used to store the models

Examples

using PlantBiophysics, MultiScaleTreeGraph, PlantGeom

# Read the file
file = joinpath(dirname(dirname(pathof(PlantBiophysics))),"test","inputs","scene","opf","coffee.opf")
mtg = read_opf(file)

# Declare our models:
models = Dict(
    "Leaf" =>
        ModelList(
            energy_balance = Monteith(),
            photosynthesis = Fvcb(),
            stomatal_conductance = Medlyn(0.03, 12.0),
            status = (d = 0.03,)
        )
)

# Checking which variables are needed for our models:
[component => to_initialize(model) for (component, model) in models]
# OK we need to initialize Rₛ, sky_fraction and the PPFD

# We can compute them directly inside the MTG from available variables:
transform!(
    mtg,
    [:Ra_PAR_f, :Ra_NIR_f] => ((x, y) -> x + y) => :Rₛ,
    :Ra_PAR_f => (x -> x * 4.57) => :PPFD,
    ignore_nothing = true
)

# Initialising all components with their corresponding models and initialisations:
init_mtg_models!(mtg, models)
# Note that this is possible only because the initialisation values are found in the MTG.
# If the initialisations are constant values between components, we can directly initialize
# them in the models definition (we initialize `:d` like this in our example).

PlantSimEngine.init_status! — Method

init_status!(object::Dict{String,ModelList};vars...)
init_status!(component::ModelList;vars...)

Intialise model variables for components with user input.

Examples

model = read_model("a-model-file.yml")
init_status!(model, Tₗ = 25.0, PPFD = 1000.0, Cₛ = 400.0, Dₗ = 1.2)

PlantSimEngine.init_variables — Method

init_variables(models...)

Intialized model variables with their default values. The variables are taken from the inputs and outputs of the models.

Examples

init_variables(Monteith())
init_variables(energy_balance = Monteith(), stomatal_conductance = Medlyn(0.03,12.0))

PlantSimEngine.init_variables_manual — Method

init_variables_manual(models...;vars...)

Return an initialisation of the model variables with given values.

Examples

init_variables_manual(status, (Tₗ = 20.0,))

PlantSimEngine.inputs — Method

inputs(model::AbstractModel)
inputs(...)

Get the inputs of one or several models.

Returns an empty tuple by default for AbstractModels (no inputs) or Missing models.

PlantSimEngine.is_initialized — Method

is_initialized(m::T) where T <: ModelList
is_initialized(m::T, models...) where T <: ModelList

Check if the variables that must be initialized are, and return true if so, and false and an information message if not.

Note

There is no way to know before-hand which process will be simulated by the user, so if you have a component with a model for each process, the variables to initialize are always the smallest subset of all, meaning it is considered the user will simulate the variables needed for other models.

Examples

leaf = ModelList(photosynthesis = Fvcb(), stomatal_conductance = Medlyn(0.03,12.0))
is_initialized(leaf)

PlantSimEngine.outputs — Method

outputs(model::AbstractModel)
outputs(...)

Get the outputs of one or several models.

Returns an empty tuple by default for AbstractModels (no outputs) or Missing models.

PlantSimEngine.pre_allocate_attr! — Method

pre_allocate_attr!(node, nsteps; attr_name = :models)

Pre-allocate the node attributes based on the status of a component model and a given number of simulation steps.

PlantSimEngine.pull_status! — Method

pull_status!(node)

Copy the status of a node's component models (e.g. the outputs of a [ModelList]@ref simulation) into the MTG attributes. This function is used when we need to compute further the simulation outputs with e.g. transform!.

Notes

Carefull, this function makes a copy, so the values are then present at two locations (can take a lot of memory space if using several plants).

Examples

using PlantBiophysics

# Read the file
mtg = read_mtg(joinpath(dirname(dirname(pathof(PlantBiophysics))),"test","inputs","scene","opf","coffee.opf"))

# Declare our models:
models = Dict(
    "Leaf" =>
        ModelList(
            energy_balance = Monteith(),
            photosynthesis = Fvcb(),
            stomatal_conductance = Medlyn(0.03, 12.0),
            status = (d = 0.03,)
        )
)

transform!(
    mtg,
    [:Ra_PAR_f, :Ra_NIR_f] => ((x, y) -> x + y) => :Rₛ,
    :Ra_PAR_f => (x -> x * 4.57) => :PPFD,
    ignore_nothing = true
)

# Initialising all components with their corresponding models and initialisations:
init_mtg_models!(mtg, models)

# Make a simulation
transform!(mtg, :models => (x -> energy_balance!(x, meteo)), ignore_nothing = true)
# Pull the simulation results into the MTG attributes:
transform!(mtg, pull_status!)
# Now the simulated variables are available from the MTG attributes field:
names(mtg)

PlantSimEngine.pull_status_one_step! — Method

pull_status_one_step!(node, step; attr_name = :models)

Copy the status of a node's ModelList (i.e. the outputs of the simulations) into the pre-allocated MTG attributes, i.e. one value per step.

See pre_allocate_attr! for the pre-allocation step.

PlantSimEngine.status — Method

status(m)
status(m::AbstractArray{<:ModelList})
status(m::AbstractDict{T,<:ModelList})

Get a ModelList status, i.e. the state of the input (and output) variables.

See also is_initialized and to_initialize

PlantSimEngine.to_initialize — Method

to_initialize(v::T, vars...) where T <: Union{Missing,AbstractModel}
to_initialize(m::T)  where T <: ModelList
to_initialize(m::DependencyTree)

Return the variables that must be initialized providing a set of models and processes. The function takes into account model coupling and only returns the variables that are needed considering that some variables that are outputs of some models are used as inputs of others.

Examples

to_initialize(photosynthesis = Fvcb(), stomatal_conductance = Medlyn(0.03,12.0))

# Or using a component directly:
leaf = ModelList(photosynthesis = Fvcb(), stomatal_conductance = Medlyn(0.03,12.0))
to_initialize(leaf)

PlantSimEngine.update_mtg_models! — Method

update_mtg_models!(mtg::MultiScaleTreeGraph.Node, i, attr_name::Symbol)

Update the mtg models initialisations by using the ith time-step. The mtg is considered fully initialized already once, so init_mtg_models! must be called before update_mtg_models!.

The values are updated only for node attributes in to_init. Those attributes must have several time-steps, i.e. indexable by 1:n time-steps.

PlantSimEngine.variables — Method

variables()

Returns a dataframe of all variables in PlantBiophysics, their description and their units.

PlantSimEngine.variables — Method

variables(model)
variables(model, models...)

Returns a tuple with the name of the variables needed by a model, or a union of those variables for several models.

Note

Each model can (and should) have a method for this function.

Examples

variables(Monteith())

variables(Monteith(), Medlyn(0.03,12.0))

See also

inputs, outputs and variables_typed

PlantSimEngine.variables_typed — Method

variables_typed(model)
variables_typed(model, models...)

Returns a named tuple with the name and the types of the variables needed by a model, or a union of those for several models.

Examples

variables_typed(Monteith())

variables_typed(Monteith(), Medlyn(0.03,12.0))

See also

inputs, outputs and variables

PlantSimEngine.vars_not_init_ — Method

vars_not_init_(st<:Status, var_names)

Get which variable is not properly initialized in the status struct.

PlantSimEngine.@gen_process_methods — Macro

@gen_process_methods(process::String, doc::String="")

This macro generate the abstract type and standard functions for a process, along with their documentation and prints out a little tutorial about how to implement a model.

The abstract process type is then used as a supertype of all models implementations for the process, and is named "Abstract<ProcessName>Model", e.g. AbstractGrowthModel for a process called growth.

The three following functions are also generated (replace "process" by your own process name):

process: a non mutating function that makes a copy of the object

process!: a mutating function that updates the object status

process!_: the actual workhorse function that does the computation, and is called by the

two previous functions under the hood. Modelers implement their own method for this function for their own model types.

The two first functions have several methods:

The base method that runs over one time-step and one object.

The method applying the computation over several objects (e.g. all leaves of a plant)

in an Array

The same method over a Dict(-alike) of objects

The method that applies the computation over several meteo time steps and

possibly several objects

A method for calling the process without any meteo (e.g. for fitting)

A method to apply the above over MTG nodes (see details)

The first argument to @gen_process_methods is the new process name, and the second is any additional documentation that should be added to the process and process! functions.

Examples

@gen_process_methods "dummy_process" "This is a dummy process that shall not be used"