示例模型

PlantSimEngine为用户提供了一些示例过程和模型。这些示例包含在子模块Examples中。要在与PlantSimEngine配套的工作环境中访问这些模型，只需使用如下子模块：

using PlantSimEngine.Examples

列表

PlantSimEngine.Examples

PlantSimEngine.Examples.AbstractCarbon_AllocationModel

PlantSimEngine.Examples.AbstractCarbon_AssimilationModel

PlantSimEngine.Examples.AbstractCarbon_BiomassModel

PlantSimEngine.Examples.AbstractCarbon_DemandModel

PlantSimEngine.Examples.AbstractDegreedaysModel

PlantSimEngine.Examples.AbstractGrowthModel

PlantSimEngine.Examples.AbstractLai_DynamicModel

PlantSimEngine.Examples.AbstractLeaf_SurfaceModel

PlantSimEngine.Examples.AbstractLight_InterceptionModel

PlantSimEngine.Examples.AbstractLight_PartitioningModel

PlantSimEngine.Examples.AbstractMaintenance_RespirationModel

PlantSimEngine.Examples.AbstractOrgan_EmergenceModel

PlantSimEngine.Examples.AbstractProcess1Model

PlantSimEngine.Examples.AbstractProcess2Model

PlantSimEngine.Examples.AbstractProcess3Model

PlantSimEngine.Examples.AbstractProcess4Model

PlantSimEngine.Examples.AbstractProcess5Model

PlantSimEngine.Examples.AbstractProcess6Model

PlantSimEngine.Examples.AbstractProcess7Model

PlantSimEngine.Examples.AbstractSoil_WaterModel

PlantSimEngine.Examples.Beer

PlantSimEngine.Examples.Process1Model

PlantSimEngine.Examples.Process2Model

PlantSimEngine.Examples.Process3Model

PlantSimEngine.Examples.Process4Model

PlantSimEngine.Examples.Process5Model

PlantSimEngine.Examples.Process6Model

PlantSimEngine.Examples.Process7Model

PlantSimEngine.Examples.ToyAssimGrowthModel

PlantSimEngine.Examples.ToyAssimModel

PlantSimEngine.Examples.ToyCAllocationModel

PlantSimEngine.Examples.ToyCBiomassModel

PlantSimEngine.Examples.ToyCDemandModel

PlantSimEngine.Examples.ToyDegreeDaysCumulModel

PlantSimEngine.Examples.ToyInternodeEmergence

PlantSimEngine.Examples.ToyLAIModel

PlantSimEngine.Examples.ToyLAIfromLeafAreaModel

PlantSimEngine.Examples.ToyLeafSurfaceModel

PlantSimEngine.Examples.ToyLightPartitioningModel

PlantSimEngine.Examples.ToyMaintenanceRespirationModel

PlantSimEngine.Examples.ToyPlantLeafSurfaceModel

PlantSimEngine.Examples.ToyPlantRmModel

PlantSimEngine.Examples.ToyRUEGrowthModel

PlantSimEngine.Examples.ToySoilWaterModel

PlantSimEngine.Examples.import_mtg_example

PlantSimEngine.fit

PlantSimEngine.run!

详情

PlantSimEngine.Examples — Module

一个包含示例模型的子模块。

这些示例用于多尺度模型集的文档。 相关模型可在包的 examples 文件夹中找到，具体存储在以下文件：

ToyAssimModel.jl

ToyCDemandModel.jl

ToyCAllocationModel.jl

ToySoilModel.jl

示例

using PlantSimEngine
using PlantSimEngine.Examples
ToyAssimModel()

PlantSimEngine.Examples.AbstractCarbon_AllocationModel — Type

carbon_allocation 过程的抽象模型类型。

所有模拟 carbon_allocation 过程的模型都必须为此类型的子类型，例如: struct MyCarbon_AllocationModel <: AbstractCarbon_AllocationModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractCarbon_AllocationModel)

PlantSimEngine.Examples.AbstractCarbon_AssimilationModel — Type

carbon_assimilation 过程的抽象模型类型。

所有模拟 carbon_assimilation 过程的模型都必须为此类型的子类型，例如: struct MyCarbon_AssimilationModel <: AbstractCarbon_AssimilationModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractCarbon_AssimilationModel)

PlantSimEngine.Examples.AbstractCarbon_BiomassModel — Type

carbon_biomass 过程的抽象模型类型。

所有模拟 carbon_biomass 过程的模型都必须为此类型的子类型，例如: struct MyCarbon_BiomassModel <: AbstractCarbon_BiomassModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractCarbon_BiomassModel)

PlantSimEngine.Examples.AbstractCarbon_DemandModel — Type

carbon_demand 过程的抽象模型类型。

所有模拟 carbon_demand 过程的模型都必须为此类型的子类型，例如: struct MyCarbon_DemandModel <: AbstractCarbon_DemandModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractCarbon_DemandModel)

PlantSimEngine.Examples.AbstractDegreedaysModel — Type

Degreedays 过程的抽象模型类型。

所有模拟 Degreedays 过程的模型都必须为此类型的子类型，例如: struct MyDegreedaysModel <: AbstractDegreedaysModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractDegreedaysModel)

PlantSimEngine.Examples.AbstractGrowthModel — Type

growth 过程的抽象模型类型。

所有模拟 growth 过程的模型都必须为此类型的子类型，例如: struct MyGrowthModel <: AbstractGrowthModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractGrowthModel)

PlantSimEngine.Examples.AbstractLai_DynamicModel — Type

LAI_Dynamic 过程的抽象模型类型。

所有模拟 LAI_Dynamic 过程的模型都必须为此类型的子类型，例如: struct MyLai_DynamicModel <: AbstractLai_DynamicModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractLai_DynamicModel)

PlantSimEngine.Examples.AbstractLeaf_SurfaceModel — Type

leaf_surface 过程的抽象模型类型。

所有模拟 leaf_surface 过程的模型都必须为此类型的子类型，例如: struct MyLeaf_SurfaceModel <: AbstractLeaf_SurfaceModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractLeaf_SurfaceModel)

PlantSimEngine.Examples.AbstractLight_InterceptionModel — Type

light_interception 过程的抽象模型类型。

所有模拟 light_interception 过程的模型都必须为此类型的子类型，例如: struct MyLight_InterceptionModel <: AbstractLight_InterceptionModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractLight_InterceptionModel)

PlantSimEngine.Examples.AbstractLight_PartitioningModel — Type

light_partitioning 过程的抽象模型类型。

所有模拟 light_partitioning 过程的模型都必须为此类型的子类型，例如: struct MyLight_PartitioningModel <: AbstractLight_PartitioningModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractLight_PartitioningModel)

PlantSimEngine.Examples.AbstractMaintenance_RespirationModel — Type

maintenance_respiration 过程的抽象模型类型。

所有模拟 maintenance_respiration 过程的模型都必须为此类型的子类型，例如: struct MyMaintenance_RespirationModel <: AbstractMaintenance_RespirationModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractMaintenance_RespirationModel)

PlantSimEngine.Examples.AbstractOrgan_EmergenceModel — Type

organ_emergence 过程的抽象模型类型。

所有模拟 organ_emergence 过程的模型都必须为此类型的子类型，例如: struct MyOrgan_EmergenceModel <: AbstractOrgan_EmergenceModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractOrgan_EmergenceModel)

PlantSimEngine.Examples.AbstractProcess1Model — Type

process1 过程的抽象模型类型。

所有模拟 process1 过程的模型都必须为此类型的子类型，例如: struct MyProcess1Model <: AbstractProcess1Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess1Model)

PlantSimEngine.Examples.AbstractProcess2Model — Type

process2 过程的抽象模型类型。

所有模拟 process2 过程的模型都必须为此类型的子类型，例如: struct MyProcess2Model <: AbstractProcess2Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess2Model)

PlantSimEngine.Examples.AbstractProcess3Model — Type

process3 过程的抽象模型类型。

所有模拟 process3 过程的模型都必须为此类型的子类型，例如: struct MyProcess3Model <: AbstractProcess3Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess3Model)

PlantSimEngine.Examples.AbstractProcess4Model — Type

process4 过程的抽象模型类型。

所有模拟 process4 过程的模型都必须为此类型的子类型，例如: struct MyProcess4Model <: AbstractProcess4Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess4Model)

PlantSimEngine.Examples.AbstractProcess5Model — Type

process5 过程的抽象模型类型。

所有模拟 process5 过程的模型都必须为此类型的子类型，例如: struct MyProcess5Model <: AbstractProcess5Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess5Model)

PlantSimEngine.Examples.AbstractProcess6Model — Type

process6 过程的抽象模型类型。

所有模拟 process6 过程的模型都必须为此类型的子类型，例如: struct MyProcess6Model <: AbstractProcess6Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess6Model)

PlantSimEngine.Examples.AbstractProcess7Model — Type

process7 过程的抽象模型类型。

所有模拟 process7 过程的模型都必须为此类型的子类型，例如: struct MyProcess7Model <: AbstractProcess7Model end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractProcess7Model)

PlantSimEngine.Examples.AbstractSoil_WaterModel — Type

soil_water 过程的抽象模型类型。

所有模拟 soil_water 过程的模型都必须为此类型的子类型，例如: struct MySoil_WaterModel <: AbstractSoil_WaterModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractSoil_WaterModel)

PlantSimEngine.Examples.Beer — Type

Beer(k)

Beer-Lambert law for light interception.

Required inputs: LAI in m² m⁻². Required meteorology data: Ri_PAR_f, the incident flux of atmospheric radiation in the PAR, in W m[soil]⁻² (== J m[soil]⁻² s⁻¹).

Output: aPPFD, the absorbed Photosynthetic Photon Flux Density in μmol[PAR] m[leaf]⁻² s⁻¹.

PlantSimEngine.Examples.Process1Model — Type

Process1Model(a)

A dummy model implementing a "process1" process for testing purposes.

PlantSimEngine.Examples.Process2Model — Type

Process2Model()

A dummy model implementing a "process2" process for testing purposes.

PlantSimEngine.Examples.Process3Model — Type

Process3Model()

A dummy model implementing a "process3" process for testing purposes.

PlantSimEngine.Examples.Process4Model — Type

Process4Model()

A dummy model implementing a "process4" process for testing purposes. It computes the inputs needed for the coupled processes 1-2-3.

PlantSimEngine.Examples.Process5Model — Type

Process5Model()

A dummy model implementing a "process5" process for testing purposes. It needs the outputs from the coupled processes 1-2-3.

PlantSimEngine.Examples.Process6Model — Type

Process6Model()

A dummy model implementing a "process6" process for testing purposes. It needs the outputs from the coupled processes 1-2-3, but also from process 7 that is itself independant.

PlantSimEngine.Examples.Process7Model — Type

Process7Model()

A dummy model implementing a "process7" process for testing purposes. It is independent (needs :var0 only as for Process4Model), but its outputs are used by Process6Model, so it is a soft-coupling.

PlantSimEngine.Examples.ToyAssimGrowthModel — Type

ToyAssimGrowthModel(Rm_factor, Rg_cost)
ToyAssimGrowthModel(; LUE=0.2, Rm_factor = 0.5, Rg_cost = 1.2)

Computes the biomass growth of a plant.

Arguments

LUE=0.2: the light use efficiency, in gC mol[PAR]⁻¹

Rm_factor=0.5: the fraction of assimilation that goes into maintenance respiration

Rg_cost=1.2: the cost of growth maintenance, in gram of carbon biomass per gram of assimilate

Inputs

aPPFD: the absorbed photosynthetic photon flux density, in mol[PAR] m⁻² time-step⁻¹

Outputs

carbon_assimilation: the assimilation, in gC m⁻² time-step⁻¹

Rm: the maintenance respiration, in gC m⁻² time-step⁻¹

Rg: the growth respiration, in gC m⁻² time-step⁻¹

biomass_increment: the daily biomass increment, in gC m⁻² time-step⁻¹

biomass: the plant biomass, in gC m⁻² time-step⁻¹

PlantSimEngine.Examples.ToyAssimModel — Type

ToyAssimModel(LUE)

Computes the assimilation of a plant (= photosynthesis).

Arguments

LUE=0.2: the light use efficiency, in gC mol[PAR]⁻¹

Inputs

aPPFD: the absorbed photosynthetic photon flux density, in mol[PAR] m⁻² time-step⁻¹

soil_water_content: the soil water content, in %

Outputs

carbon_assimilation: the assimilation or photosynthesis, also sometimes denoted A, in gC m⁻² time-step⁻¹

Details

The assimilation is computed as the product of the absorbed photosynthetic photon flux density (aPPFD) and the light use efficiency (LUE), so the units of the assimilation usually are in gC m⁻² time-step⁻¹, but they could be in another spatial or temporal unit depending on the unit of aPPFD, e.g. if aPPFD is in mol[PAR] plant⁻¹ time-step⁻¹, the assimilation will be in gC plant⁻¹ time-step⁻¹.

PlantSimEngine.Examples.ToyCAllocationModel — Type

ToyCAllocationModel()

Computes the carbon allocation to each organ of a plant based on the plant total carbon offer and individual organ demand. This model should be used at the plant scale, because it first computes the carbon availaible for allocation as the minimum between the total demand (sum of organs' demand) and total carbon offer (sum of organs' assimilation - total maintenance respiration), and then allocates the carbon relative to each organ's demand.

Inputs

carbon_assimilation: a vector of the assimilation of all photosynthetic organs, usually in gC m⁻² time-step⁻¹

Rm: the maintenance respiration of the plant, usually in gC m⁻² time-step⁻¹

carbon_demand: a vector of the carbon demand of the organs, usually in gC m⁻² time-step⁻¹

Outputs

carbon_assimilation: the carbon assimilation, usually in gC m⁻² time-step⁻¹

Details

The units usually are in gC m⁻² time-step⁻¹, but they could be in another spatial or temporal unit depending on the unit of the inputs, e.g. in gC plant⁻¹ time-step⁻¹.

PlantSimEngine.Examples.ToyCBiomassModel — Type

ToyCBiomassModel(construction_cost)

Computes the carbon biomass of an organ based on the carbon allocation and construction cost.

Arguments

construction_cost: the construction cost of the organ, usually in gC gC⁻¹. Should be understood as the amount of carbon needed to build 1g of carbon biomass.

Inputs

carbon_allocation: the carbon allocation to the organ for the time-step, usually in gC m⁻² time-step⁻¹

Outputs

carbon_biomass_increment: the increment of carbon biomass, usually in gC time-step⁻¹

carbon_biomass: the carbon biomass, usually in gC

growth_respiration: the growth respiration, usually in gC time-step⁻¹

PlantSimEngine.Examples.ToyCDemandModel — Type

ToyCDemandModel(optimal_biomass, development_duration)
ToyCDemandModel(; optimal_biomass, development_duration)

Computes the carbon demand of an organ depending on its biomass under optimal conditions and the duration of its development in degree days. The model assumes that the carbon demand is linear througout the duration of the development.

Arguments

optimal_biomass: the biomass of the organ under optimal conditions, in gC

development_duration: the duration of the development of the organ, in degree days

Inputs

TT: the thermal time, in degree days

Outputs

carbon_demand: the carbon demand, in gC

PlantSimEngine.Examples.ToyDegreeDaysCumulModel — Type

ToyDegreeDaysCumulModel(;init_TT=0.0, T_base=10.0, T_max=43.0)

Computes the thermal time in degree days and cumulated degree-days based on the average daily temperature (T), the initial cumulated degree days, the base temperature below which there is no growth, and the maximum temperature for growh.

PlantSimEngine.Examples.ToyInternodeEmergence — Type

ToyInternodeEmergence(;init_TT=0.0, TT_emergence = 300)

Computes the organ emergence based on cumulated thermal time since last event.

PlantSimEngine.Examples.ToyLAIModel — Type

ToyLAIModel(;max_lai=8.0, dd_incslope=800, inc_slope=110, dd_decslope=1500, dec_slope=20)

Computes the Leaf Area Index (LAI) based on a sigmoid function of thermal time.

Arguments

max_lai: the maximum LAI value

dd_incslope: the thermal time at which the LAI starts to increase

inc_slope: the slope of the increase

dd_decslope: the thermal time at which the LAI starts to decrease

dec_slope: the slope of the decrease

Inputs

TT_cu: the cumulated thermal time since the beginning of the simulation, usually in °C days

Outputs

LAI: the Leaf Area Index, usually in m² m⁻²

PlantSimEngine.Examples.ToyLAIfromLeafAreaModel — Type

ToyLAIfromLeafAreaModel()

Computes the Leaf Area Index (LAI) of the scene based on the plants leaf area.

Arguments

scene_area: the area of the scene, usually in m²

Inputs

surface: a vector of plant leaf surfaces, usually in m²

Outputs

LAI: the Leaf Area Index of the scene, usually in m² m⁻²

total_surface: the total surface of the plants, usually in m²

PlantSimEngine.Examples.ToyLeafSurfaceModel — Type

ToyLeafSurfaceModel(SLA)

Computes the individual leaf surface from its biomass using the SLA.

Arguments

SLA: the specific leaf area, usually in m² gC⁻¹. Should be understood as the surface area of a leaf per unit of carbon biomass.

Values typically range from 0.002 to 0.027 m² gC⁻¹.

Inputs

carbon_biomass: the carbon biomass of the leaf, usually in gC

Outputs

surface: the leaf surface, usually in m²

PlantSimEngine.Examples.ToyLightPartitioningModel — Type

ToyLightPartitioningModel()

Computes the light partitioning based on relative surface.

Inputs

aPPFD: the absorbed photosynthetic photon flux density at the larger scale (e.g. scene), in mol[PAR] m⁻² time-step⁻¹

Outputs

aPPFD: the assimilation or photosynthesis, also sometimes denoted A, in gC time-step⁻¹

Details

PlantSimEngine.Examples.ToyMaintenanceRespirationModel — Type

RmQ10FixedN(Q10, Rm_base, T_ref, P_alive, nitrogen_content)

Maintenance respiration based on a Q10 computation with fixed nitrogen values and proportion of living cells in the organs.

Arguments

Q10: Q10 factor (values should usually range between: 1.5 - 2.5, with 2.1 being the most common value)

Rm_base: Base maintenance respiration (gC gDM⁻¹ time-step⁻¹). Should be around 0.06.

T_ref: Reference temperature at which Q10 was measured (usually around 25.0°C)

P_alive: proportion of living cells in the organ

nitrogen_content: nitrogen content of the organ (gN gC⁻¹)

Inputs

carbon_biomass: the carbon biomass of the organ in gC

PlantSimEngine.Examples.ToyPlantLeafSurfaceModel — Type

ToyPlantLeafSurfaceModel()

Computes the leaf surface at plant scale by summing the individual leaf surfaces.

Inputs

leaf_surfaces: a vector of leaf surfaces, usually in m²

Outputs

surface: the leaf surface at plant scale, usually in m²

PlantSimEngine.Examples.ToyPlantRmModel — Type

ToyPlantRmModel()

Total plant maintenance respiration based on the sum of Rm_organs, the maintenance respiration of the organs.

Intputs

Rm_organs: a vector of maintenance respiration from all organs in the plant in gC time-step⁻¹

Outputs

Rm: the total plant maintenance respiration in gC time-step⁻¹

PlantSimEngine.Examples.ToyRUEGrowthModel — Type

ToyRUEGrowthModel(efficiency)

Computes the carbon biomass increment of a plant based on the radiation use efficiency principle.

Arguments

efficiency: the radiation use efficiency, in gC[biomass] mol[PAR]⁻¹

Inputs

aPPFD: the absorbed photosynthetic photon flux density, in mol[PAR] m⁻² time-step⁻¹

Outputs

biomass_increment: the daily biomass increment, in gC[biomass] m⁻² time-step⁻¹

biomass: the plant biomass, in gC[biomass] m⁻² time-step⁻¹

PlantSimEngine.Examples.ToySoilWaterModel — Type

ToySoilWaterModel(values=[0.5])

A toy model to compute the soil water content. The model simply take a random value in the values range using rand.

Outputs

soil_water_content: the soil water content (%).

Arguments

values: a range of soil_water_content values to sample from. Can be a vector of values [0.5,0.6] or a range 0.1:0.1:1.0. Default is [0.5].

PlantSimEngine.Examples.import_mtg_example — Method

import_mtg_example()

返回一个带有场景、土壤和一株带有两个节间和两个叶片的植物的多尺度树图（MTG）示例。

示例

julia> using PlantSimEngine.Examples

julia> import_mtg_example()
/ 1: Scene
├─ / 2: Soil
└─ + 3: Plant
   └─ / 4: Internode
      ├─ + 5: Leaf
      └─ < 6: Internode
         └─ + 7: Leaf

PlantSimEngine.fit — Method

fit(::Type{Beer}, df; J_to_umol=PlantMeteo.Constants().J_to_umol)

Compute the k parameter of the Beer-Lambert law from measurements.

Arguments

::Type{Beer}: the model type

df: a DataFrame with the following columns:

aPPFD: the measured absorbed Photosynthetic Photon Flux Density in μmol[PAR] m[leaf]⁻² s⁻¹

LAI: the measured leaf area index in m² m⁻²

Ri_PAR_f: the measured incident flux of atmospheric radiation in the PAR, in W m[soil]⁻² (== J m[soil]⁻² s⁻¹)

Examples

Import the example models defined in the Examples sub-module:

using PlantSimEngine
using PlantSimEngine.Examples

Create a model list with a Beer model, and fit it to the data:

m = ModelList(Beer(0.6), status=(LAI=2.0,))
meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_f=300.0)
run!(m, meteo)
df = DataFrame(aPPFD=m[:aPPFD][1], LAI=m.status.LAI[1], Ri_PAR_f=meteo.Ri_PAR_f[1])
fit(Beer, df)

PlantSimEngine.run! — Function

run!(::Beer, object, meteo, constants=Constants(), extra=nothing)

Computes the photosynthetic photon flux density (aPPFD, µmol m⁻² s⁻¹) absorbed by an object using the incoming PAR radiation flux (Ri_PAR_f, W m⁻²) and the Beer-Lambert law of light extinction.

Arguments

::Beer: a Beer model, from the model list (i.e. m.light_interception)

models: A ModelList struct holding the parameters for the model with

initialisations for LAI (m² m⁻²): the leaf area index.

status: the status of the model, usually the model list status (i.e. m.status)

meteo: meteorology structure, see Atmosphere

constants = PlantMeteo.Constants(): physical constants. See PlantMeteo.Constants for more details

extra = nothing: extra arguments, not used here.

Examples

m = ModelList(Beer(0.5), status=(LAI=2.0,))

meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_q=300.0)

run!(m, meteo)

m[:aPPFD]