API - 内部函数

未导出成员

PlantSimEngine 中的私有函数、类型或常量。这些内容均未导出，因此需要通过 PlantSimEngine. 前缀来访问（例如 PlantSimEngine.DataFormat）。大多数属于开发者代码，但其中有些内容对希望深入探索或需更精细控制模拟参数的用户也有用（未来文档版本可能会将这些类别分拆到不同页面以提升清晰度）。

索引

DataFrames.DataFrame

DataFrames.DataFrame

PlantSimEngine.AbstractBasic_Current_TimestepModel

PlantSimEngine.AbstractBasic_Next_TimestepModel

PlantSimEngine.AbstractNodeMapping

PlantSimEngine.DataFormat

PlantSimEngine.DependencyGraph

PlantSimEngine.DependencyTrait

PlantSimEngine.GraphSimulation

PlantSimEngine.MappedVar

PlantSimEngine.MultiNodeMapping

PlantSimEngine.ObjectDependencyTrait

PlantSimEngine.RefVariable

PlantSimEngine.RefVector

PlantSimEngine.SelfNodeMapping

PlantSimEngine.SingleNodeMapping

PlantSimEngine.TimeStepDependencyTrait

PlantSimEngine.UninitializedVar

Base.copy

Base.copy

Base.copy

Base.getindex

PlantSimEngine.add_mapped_variables_with_outputs_as_inputs!

PlantSimEngine.add_model_vars

PlantSimEngine.check_dimensions

PlantSimEngine.convert_reference_values!

PlantSimEngine.convert_vars

PlantSimEngine.convert_vars!

PlantSimEngine.convert_vars!

PlantSimEngine.default_variables_from_mapping

PlantSimEngine.diff_vars

PlantSimEngine.draw_guide

PlantSimEngine.draw_panel

PlantSimEngine.drop_process

PlantSimEngine.flatten_vars

PlantSimEngine.get_mapped_variables

PlantSimEngine.get_model_nodes

PlantSimEngine.get_models

PlantSimEngine.get_multiscale_default_value

PlantSimEngine.get_nsteps

PlantSimEngine.get_status

PlantSimEngine.hard_dependencies

PlantSimEngine.homogeneous_ts_kwargs

PlantSimEngine.homogeneous_ts_kwargs

PlantSimEngine.init_node_status!

PlantSimEngine.init_simulation

PlantSimEngine.init_statuses

PlantSimEngine.init_variables_manual

PlantSimEngine.initialise_all_as_hard_dependency_node

PlantSimEngine.is_graph_cyclic

PlantSimEngine.mapped_variables

PlantSimEngine.mapped_variables_no_outputs_from_other_scale

PlantSimEngine.model_

PlantSimEngine.object_parallelizable

PlantSimEngine.parallelizable

PlantSimEngine.ref_var

PlantSimEngine.reverse_mapping

PlantSimEngine.save_results!

PlantSimEngine.search_inputs_in_multiscale_output

PlantSimEngine.search_inputs_in_output

PlantSimEngine.set_variables_at_timestep!

PlantSimEngine.soft_dependencies

PlantSimEngine.status_from_template

PlantSimEngine.timestep_parallelizable

PlantSimEngine.transform_single_node_mapped_variables_as_self_node_output!

PlantSimEngine.traverse_dependency_graph

PlantSimEngine.traverse_dependency_graph!

PlantSimEngine.traverse_dependency_graph!

PlantSimEngine.variables_multiscale

PlantSimEngine.variables_outputs_from_other_scale

PlantSimEngine.variables_typed

PlantSimEngine.vars_not_init_

API 文档

DataFrames.DataFrame — Method

DataFrame(components <: AbstractArray{<:ModelList})
DataFrame(components <: AbstractDict{N,<:ModelList})

从ModelList（或其数组/字典）的状态中提取数据为一个DataFrame。

示例

using PlantSimEngine
using DataFrames

# 创建一个 ModelList
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

# 转换为 DataFrame
df = DataFrame(models)

# 创建一个包含 ModelList 的字典
models = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    ),
    "InterNode" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    )
)

# 转换为 DataFrame
df = DataFrame(models)

DataFrames.DataFrame — Method

DataFrame(components::ModelList{T,S}) where {T,S<:Status}

用于只有一个时间步的 ModelList 模型的 DataFrame 实现。

PlantSimEngine.AbstractBasic_Current_TimestepModel — Type

basic_current_timestep 过程的抽象模型类型。

所有模拟 basic_current_timestep 过程的模型都必须为此类型的子类型，例如: struct MyBasic_Current_TimestepModel <: AbstractBasic_Current_TimestepModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractBasic_Current_TimestepModel)

PlantSimEngine.AbstractBasic_Next_TimestepModel — Type

basic_next_timestep 过程的抽象模型类型。

所有模拟 basic_next_timestep 过程的模型都必须为此类型的子类型，例如: struct MyBasic_Next_TimestepModel <: AbstractBasic_Next_TimestepModel end。

可通过 subtypes 查看实现该过程的所有模型：

例子

subtypes(AbstractBasic_Next_TimestepModel)

PlantSimEngine.AbstractNodeMapping — Type

AbstractNodeMapping

用于节点映射类型的抽象类型，例如单节点映射或多节点映射。

PlantSimEngine.DataFormat — Method

DataFormat(T::Type)

返回类型 T 的数据格式。数据格式用于确定如何遍历数据。支持以下数据格式：

TableAlike：数据为类似表格的对象，例如 DataFrame 或 TimeStepTable。数据通过 Tables.jl 接口按行遍历。

SingletonAlike：数据为类似单例的对象，例如 NamedTuple 或 TimeStepRow。数据通过按列遍历。

TreeAlike：数据为类似树结构的对象，例如 Node。

默认实现中，AbstractDataFrame、TimeStepTable、AbstractVector 和 Dict 返回 TableAlike；GraphSimulation 返回 TreeAlike；Status、ModelList、NamedTuple 和 TimeStepRow 返回 SingletonAlike。

Any 的默认实现会抛出异常。如果用户希望使用其他输入类型，应为新数据格式定义此 trait，例如：

PlantSimEngine.DataFormat(::Type{<:MyType}) = TableAlike()

示例

julia> using PlantSimEngine, PlantMeteo, DataFrames

julia> PlantSimEngine.DataFormat(DataFrame)
PlantSimEngine.TableAlike()

julia> PlantSimEngine.DataFormat(TimeStepTable([Status(a = 1, b = 2, c = 3)]))
PlantSimEngine.TableAlike()

julia> PlantSimEngine.DataFormat([1, 2, 3])
PlantSimEngine.TableAlike()

julia> PlantSimEngine.DataFormat(Dict(:a => 1, :b => 2))
PlantSimEngine.TableAlike()

julia> PlantSimEngine.DataFormat(Status(a = 1, b = 2, c = 3))
PlantSimEngine.SingletonAlike()

PlantSimEngine.DependencyGraph — Type

DependencyGraph{T}(roots::T, not_found::Dict{Symbol,DataType})

模型之间依赖关系的图。

参数

roots::T：图的根节点。

not_found::Dict{Symbol,DataType}：在图中未找到的模型。

PlantSimEngine.DependencyTrait — Type

DependencyTrait(T::Type)

返回关于模型 T 在计算时是否依赖于其他时间步或对象的信息。依赖特性用于判断模型能否并行化。

支持以下依赖特性：

TimeStepDependencyTrait：定义模型对于时间步是否可以并行化计算的特性。

ObjectDependencyTrait：定义模型对于对象是否可以并行化计算的特性。

PlantSimEngine.GraphSimulation — Type

GraphSimulation(graph, mapping)
GraphSimulation(graph, statuses, dependency_graph, models, outputs)

一个类型，包含了图上仿真所需的所有信息。

参数

graph：一个图实例，比如 MTG

mapping：模型映射字典

statuses：定义每个节点状态的结构体

status_templates：状态模板字典

reverse_multiscale_mapping：多尺度下其它层次的映射字典

var_need_init：需要初始化的变量字典

dependency_graph：应用于图的模型的依赖图

models：模型字典

outputs：输出字典

PlantSimEngine.MappedVar — Type

MappedVar(source_organ, variable, source_variable, source_default)

映射到其他层级的变量。

参数

source_organ: 映射所针对的器官（或多个器官）

variable: 被映射的变量名

source_variable: 来源器官中的变量名（实际计算该变量的变量名）

source_default: 变量的默认值

示例

julia> using PlantSimEngine

julia> PlantSimEngine.MappedVar(PlantSimEngine.SingleNodeMapping("Leaf"), :carbon_assimilation, :carbon_assimilation, 1.0)
PlantSimEngine.MappedVar{PlantSimEngine.SingleNodeMapping, Symbol, Symbol, Float64}(PlantSimEngine.SingleNodeMapping("Leaf"), :carbon_assimilation, :carbon_assimilation, 1.0)

PlantSimEngine.MultiNodeMapping — Type

MultiNodeMapping(scale)

用于多节点映射的类型，例如 [:carbon_assimilation => ["Leaf"],]。注意 "Leaf" 以向量形式给出， 这意味着 :carbon_assimilation 将会是从植物图中每个 "Leaf" 取得的值的向量。

PlantSimEngine.ObjectDependencyTrait — Type

ObjectDependencyTrait(::Type{T})

定义模型 T 计算时是否依赖其他对象的特性。该依赖特性用于判断模型是否可以在对象上并行化。

支持以下依赖特性：

IsObjectDependent：模型依赖其他对象计算，不能并行。

IsObjectIndependent：模型不依赖其他对象，可以并行。

所有模型默认设为对象依赖（即 IsObjectDependent）。这样设计的考虑：

这是最安全的默认值，如果用户忘记重载不会导致错误，而相反则可能出错。

对于对象独立模型，用户可以方便重载 trait。

参见

timestep_parallelizable：判断模型能否按时间步并行。

object_parallelizable：判断模型能否按对象并行。

parallelizable：判断模型是否可并行化。

TimeStepDependencyTrait：定义模型对其他时间步的依赖性。

示例

定义一个测试过程:

using PlantSimEngine

# 定义一个测试过程:
@process "TestProcess"

定义一个对象独立的模型:

struct MyModel <: AbstractTestprocessModel end

# 重载对象依赖 trait:
PlantSimEngine.ObjectDependencyTrait(::Type{MyModel}) = IsObjectIndependent()

检查模型能否按对象并行:

object_parallelizable(MyModel()) # false

定义一个对象依赖的模型:

struct MyModel2 <: AbstractTestprocessModel end

# 重载对象依赖 trait:
PlantSimEngine.ObjectDependencyTrait(::Type{MyModel2}) = IsObjectDependent()

检查模型能否按对象并行:

object_parallelizable(MyModel()) # true

PlantSimEngine.RefVariable — Type

RefVariable(reference_variable)

用于手动标记模型中的某个变量，以在相同尺度下采用另一个变量的值的结构体。 用于变量重命名，当某变量已经由一个模型计算，但被另一个名称所引用时使用。

注意：在 status 中我们并不会真正重命名变量（其他模型可能需要这个变量），而是新建一个对原有变量的引用。

PlantSimEngine.RefVector — Type

RefVector(field::Symbol, sts...)
RefVector(field::Symbol, sts::Vector{<:Status})
RefVector(v::Vector{Base.RefValue{T}})

一个引用结构体向量中特定字段的引用向量。 用于在不同尺度之间高效地传递数值。

参数

field: 要引用的结构体字段

sts...: 要引用的结构体实例

sts::Vector{<:Status}: 要引用的结构体实例向量

示例

julia> using PlantSimEngine

假设有两个 Status 结构体：

julia> status1 = Status(a = 1.0, b = 2.0, c = 3.0);

julia> status2 = Status(a = 2.0, b = 3.0, c = 4.0);

我们可以创建引用结构体 status1 和 status2 字段 a 的 RefVector：

julia> rv = PlantSimEngine.RefVector(:a, status1, status2)
2-element PlantSimEngine.RefVector{Float64}:
 1.0
 2.0

这等效于：

julia> rv = PlantSimEngine.RefVector(:a, [status1, status2])
2-element PlantSimEngine.RefVector{Float64}:
 1.0
 2.0

可以访问 RefVector 的值：

julia> rv[1]
1.0

修改 RefVector 中的值会同时修改原结构体中的值：

julia> rv[1] = 10.0
10.0

julia> status1.a
10.0

我们还可以通过引用向量创建 RefVector：

julia> vec = [Ref(1.0), Ref(2.0), Ref(3.0)]
3-element Vector{Base.RefValue{Float64}}:
 Base.RefValue{Float64}(1.0)
 Base.RefValue{Float64}(2.0)
 Base.RefValue{Float64}(3.0)

julia> rv = PlantSimEngine.RefVector(vec)
3-element PlantSimEngine.RefVector{Float64}:
 1.0
 2.0
 3.0

julia> rv[1]
1.0

PlantSimEngine.SelfNodeMapping — Type

SelfNodeMapping()

自节点映射类型，即节点映射到其自身。 用于标记将会被其他模型以标量值引用的变量。可出现于两种情况： - 该变量由其他层级计算，因此需要作为输入存在于本层级（否则不会在本层级计算） - 该变量作为输入用于其他层级，但仅以单值（标量）引用，因此需要将其作为标量引用。

PlantSimEngine.SingleNodeMapping — Type

SingleNodeMapping(scale)

用于单节点映射的类型，例如 [:soil_water_content => "Soil",]。注意 "Soil" 是以标量形式给出， 这意味着 :soil_water_content 将会是从植物图中特定的 "Soil" 节点取得的一个标量值。

PlantSimEngine.TimeStepDependencyTrait — Method

TimeStepDependencyTrait(::Type{T})

定义模型 T 计算时是否依赖其他时间步的特性。该依赖特性用于判断模型是否可以在时间步上并行化。

支持以下依赖特性：

IsTimeStepDependent：该模型的计算依赖其他时间步，不能并行执行。

IsTimeStepIndependent：该模型的计算不依赖其他时间步，可以并行执行。

所有模型默认设为时间步依赖（即 IsTimeStepDependent）。虽然大多数模型不一定如此，但这样设计有两个原因：

这是最安全的默认值——如果用户忘记重载这个 trait，不会导致错误结果，相反（设为独立）则可能出错。

对于时间步独立的模型，用户可以很方便地重载此 trait。

参见

timestep_parallelizable：判断模型能否在时间步并行。

object_parallelizable：判断模型能否在对象间并行。

parallelizable：判断模型是否可并行化。

ObjectDependencyTrait：定义模型对其他对象的依赖性。

示例

定义一个测试过程:

using PlantSimEngine

# 定义一个测试过程:
@process "TestProcess"

定义一个时间步独立的模型:

struct MyModel <: AbstractTestprocessModel end

# 重载时间步依赖 trait:
PlantSimEngine.TimeStepDependencyTrait(::Type{MyModel}) = IsTimeStepIndependent()

检查模型能否按时间步并行:

timestep_parallelizable(MyModel()) # false

定义一个时间步依赖的模型:

struct MyModel2 <: AbstractTestprocessModel end

# 重载时间步依赖 trait:
PlantSimEngine.TimeStepDependencyTrait(::Type{MyModel2}) = IsTimeStepDependent()

检查模型能否按时间步并行:

timestep_parallelizable(MyModel()) # true

PlantSimEngine.UninitializedVar — Type

UninitializedVar(variable, value)

尚未初始化的变量，赋予一个名称和一个默认值。

Base.copy — Method

Base.copy(l::AbstractArray{<:ModelList})

复制 ModelList 的数组类结构

Base.copy — Method

Base.copy(l::AbstractDict{N,<:ModelList} where N)

复制字典类 ModelList

Base.copy — Method

Base.copy(l::ModelList)
Base.copy(l::ModelList, status)

复制一个 ModelList，可选替换新的 status。

示例

using PlantSimEngine

# 包含示例过程和模型：
using PlantSimEngine.Examples;

# 创建模型列表：
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

# 复制模型列表：
ml2 = copy(models)

# 使用新 status 复制模型列表：
ml3 = copy(models, TimeStepTable([Status(var1=20.0, var2=0.5))])

Base.getindex — Method

getindex(component<:ModelList, key::Symbol)
getindex(component<:ModelList, key)

对组件模型结构进行索引: - 当参数为整数时，会返回第 i 个时间步的状态 - 其他（Symbol、String）则会返回状态中的对应变量

示例

using PlantSimEngine

lm = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status = (var1=[15.0, 16.0], var2=0.3)
);

lm[:var1] # 返回 Tₗ 变量的值
lm[2]  # 返回第二个时间步的状态
lm[2][:var1] # 返回第二个时间步的 Tₗ 变量的值
lm[:var1][2] # 等价于上面

# 输出
16.0

PlantSimEngine.add_mapped_variables_with_outputs_as_inputs! — Method

add_mapped_variables_with_outputs_as_inputs!(mapped_vars)

将计算于一个尺度并写入到另一个尺度的变量添加进映射字典。

PlantSimEngine.add_model_vars — Method

add_model_vars(x, models, type_promotion)

检测 x 中哪些变量还未初始化（根据一组 models 及其所需模拟变量）。如有变量未初始化，将其初始化为默认值。

本函数需根据不同类型的 x 实现。默认方法适用于任意 Tables.jl 兼容结构和 NamedTuples。

注意，该函数会在变量未齐全时复制输入 x。

PlantSimEngine.check_dimensions — Method

check_dimensions(component,weather)
check_dimensions(status,weather)

检查组件状态（或直接检查状态）和天气数据长度是否一致，或它们是否可以被复用（其中一个的长度为1）。

示例

using PlantSimEngine, PlantMeteo

# 引入包含虚拟过程和模型的示例脚本：
using PlantSimEngine.Examples

# 创建一个虚拟天气：
w = Atmosphere(T = 20.0, Rh = 0.5, Wind = 1.0)

# 创建一个虚拟组件：
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=[15.0, 16.0], var2=0.3)
)

# 检查时间步数是否兼容（此处兼容，返回nothing）：
PlantSimEngine.check_dimensions(models, w) 

# 创建一个包含3个时间步的虚拟天气：
w = Weather([
    Atmosphere(T = 20.0, Rh = 0.5, Wind = 1.0),
    Atmosphere(T = 25.0, Rh = 0.5, Wind = 1.0),
    Atmosphere(T = 30.0, Rh = 0.5, Wind = 1.0)
])

# 检查时间步数是否兼容（此处不兼容，会抛出错误）：
PlantSimEngine.check_dimensions(models, w)

# output
ERROR: DimensionMismatch: Component status has a vector variable : var1 implying multiple timesteps but weather data only provides a single timestep.

PlantSimEngine.convert_reference_values! — Method

convert_reference_values!(mapped_vars::Dict{String,Dict{Symbol,Any}})

将MappedVar{SelfNodeMapping}或MappedVar{SingleNodeMapping}变量转为参考同一变量值的RefValue； 将MappedVar{MultiNodeMapping}转为RefVector，引用源器官的变量值。

PlantSimEngine.convert_vars — Function

convert_vars(ref_vars, type_promotion::Dict{DataType,DataType})
convert_vars(ref_vars, type_promotion::Nothing)
convert_vars!(ref_vars::Dict{Symbol}, type_promotion::Dict{DataType,DataType})
convert_vars!(ref_vars::Dict{Symbol}, type_promotion::Nothing)

将 status 变量转换为 type_promotion 字典中指定的类型。 注：带 ! 的变异版仅适用于变量字典。

示例

若需将所有 Real 类型变量转为 Float32，可使用：

using PlantSimEngine

# 包含示例过程和模型：
using PlantSimEngine.Examples;

ref_vars = init_variables(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
)
type_promotion = Dict(Real => Float32)

PlantSimEngine.convert_vars(type_promotion, ref_vars.process3)

PlantSimEngine.convert_vars! — Function

convert_vars(ref_vars, type_promotion::Dict{DataType,DataType})
convert_vars(ref_vars, type_promotion::Nothing)
convert_vars!(ref_vars::Dict{Symbol}, type_promotion::Dict{DataType,DataType})
convert_vars!(ref_vars::Dict{Symbol}, type_promotion::Nothing)

将 status 变量转换为 type_promotion 字典中指定的类型。 注：带 ! 的变异版仅适用于变量字典。

示例

若需将所有 Real 类型变量转为 Float32，可使用：

using PlantSimEngine

# 包含示例过程和模型：
using PlantSimEngine.Examples;

ref_vars = init_variables(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
)
type_promotion = Dict(Real => Float32)

PlantSimEngine.convert_vars(type_promotion, ref_vars.process3)

PlantSimEngine.convert_vars! — Method

convert_vars!(mapped_vars::Dict{String,Dict{String,Any}}, type_promotion)

使用 type_promotion 字典将映射（mapped_vars）中的变量类型进行转换。

mapped_vars 应为器官名称到变量字典的字典，变量名为 Symbol，变量值为值。

PlantSimEngine.default_variables_from_mapping — Function

default_variables_from_mapping(mapped_vars, verbose=true)

递归从映射关系中查找原始映射值，获取映射变量的默认值。

参数

mapped_vars::Dict{String,Dict{Symbol,Any}}: 每个器官的映射变量。

verbose::Bool: 是否打印在mapping中查找默认值的堆栈追踪。

PlantSimEngine.diff_vars — Method

diff_vars(x, y)

返回 x 和 y 中值不同的变量名称。

PlantSimEngine.draw_guide — Method

draw_guide(h, w, prefix, isleaf, guides)

为依赖图的一个节点绘制连线引导

PlantSimEngine.draw_panel — Function

draw_panel(node, graph, prefix, dep_graph_guides, parent; title="Soft-coupled model")

为所有依赖项绘制面板

PlantSimEngine.drop_process — Method

drop_process(proc_vars, process)

返回将 NamedTuple proc_vars 中 process process 删除后的新 NamedTuple。

参数

proc_vars::NamedTuple: 要删除进程的 NamedTuple。

process::Symbol: 需要从 NamedTuple proc_vars 中删除的进程。

返回

删除了 process 后的 NamedTuple。

示例

julia> drop_process((a = 1, b = 2, c = 3), :b)
(a = 1, c = 3)

julia> drop_process((a = 1, b = 2, c = 3), (:a, :c))
(b = 2,)

PlantSimEngine.flatten_vars — Method

flatten_vars(vars)

返回 vars 字典中的变量集合。

参数

vars::Dict{Symbol, Tuple{Symbol, Vararg{Symbol}}}: 进程到变量内容的字典。

返回

vars 字典中的变量集合。

示例

julia> flatten_vars(Dict(:process1 => (:var1, :var2), :process2 => (:var3, :var4)))
Set{Symbol} with 4 elements:
  :var4
  :var3
  :var2
  :var1

julia> flatten_vars([:process1 => (var1 = -Inf, var2 = -Inf), :process2 => (var3 = -Inf, var4 = -Inf)])
(var2 = -Inf, var4 = -Inf, var3 = -Inf, var1 = -Inf)

PlantSimEngine.get_mapped_variables — Method

get_mapped_variables(m)

获取模型映射字典中的变量映射。

参数

m::Dict{String,Any}: 一个模型映射的字典

返回由符号和字符串或字符串向量组成的Pair的向量。

示例

参见get_models中的示例。

PlantSimEngine.get_model_nodes — Method

get_model_nodes(dep_graph::DependencyGraph, model)

获取依赖图中实现某种模型类型的节点。

参数

dep_graph::DependencyGraph：依赖关系图。

model：要查找的模型类型。

返回

实现该模型类型的节点数组。

示例

PlantSimEngine.get_model_nodes(dependency_graph, Beer)

PlantSimEngine.get_models — Method

get_models(m)

获取模型映射字典中的模型。

参数

m::Dict{String,Any}: 一个模型映射的字典

返回一个模型的向量

示例

julia> using PlantSimEngine;

导入示例模型（可以在本包的examples文件夹或Examples子模块中找到）:

julia> using PlantSimEngine.Examples;

如果只提供一个MultiScaleModel，将获得其模型，结果为单元素向量：

julia> models = MultiScaleModel( model=ToyCAllocationModel(), mapped_variables=[ :carbon_assimilation => ["Leaf"], :carbon_demand => ["Leaf", "Internode"], :carbon_allocation => ["Leaf", "Internode"] ], );

julia> PlantSimEngine.get_models(models)
1-element Vector{ToyCAllocationModel}:
 ToyCAllocationModel()

如果提供一个模型元组，将分别得到每个模型，返回为向量：

julia> models2 = (  MultiScaleModel( model=ToyAssimModel(), mapped_variables=[:soil_water_content => "Soil",], ), ToyCDemandModel(optimal_biomass=10.0, development_duration=200.0), Status(aPPFD=1300.0, TT=10.0), );

注意此处在映射时提供的是"Soil"而不是["Soil"]，因为此处只需要一个值。

julia> PlantSimEngine.get_models(models2)
2-element Vector{AbstractModel}:
 ToyAssimModel{Float64}(0.2)
 ToyCDemandModel{Float64}(10.0, 200.0)

PlantSimEngine.get_multiscale_default_value — Function

get_multiscale_default_value(mapped_vars, val, mapping_stacktrace=[])

从映射关系获取变量的默认值。

参数

mapped_vars::Dict{String,Dict{Symbol,Any}}: 每个器官的映射变量。

val::Any: 需要获取默认值的变量。

mapping_stacktrace::Vector{Any}: 在向上查找映射值时的堆栈追踪。

PlantSimEngine.get_nsteps — Method

get_nsteps(t)

获取对象的时间步数量。

PlantSimEngine.get_status — Method

get_status(m)

获取模型映射字典中的状态。

参数

m::Dict{String,Any}: 一个模型映射的字典

返回Status或nothing。

示例

参见get_models中的示例。

PlantSimEngine.hard_dependencies — Method

hard_dependencies(models; verbose::Bool=true)
hard_dependencies(mapping::Dict{String,T}; verbose::Bool=true)

计算模型之间的硬依赖关系。

PlantSimEngine.homogeneous_ts_kwargs — Method

homogeneous_ts_kwargs(kwargs)

默认情况下，此函数直接返回其参数。

PlantSimEngine.homogeneous_ts_kwargs — Method

kwargs_to_timestep(kwargs::NamedTuple{N,T}) where {N,T}

将各变量为（可选）向量的 NamedTuple，转换为每个时间步为一个 NamedTuple 的向量。 可用于对某一变量在所有时间步赋常值。

示例

PlantSimEngine.homogeneous_ts_kwargs((Tₗ=[25.0, 26.0], aPPFD=1000.0))

PlantSimEngine.init_node_status! — Function

init_node_status!(
    node, 
    statuses, 
    mapped_vars, 
    reverse_multiscale_mapping,
    vars_need_init=Dict{String,Any}(),
    type_promotion=nothing;
    check=true,
    attribute_name=:plantsimengine_status)
)

初始化植物图节点的状态，考虑多尺度映射，并将状态添加到statuses字典中。

参数

node：要初始化的节点

statuses：按节点类型分的状态字典

mapped_vars：每个节点类型的状态模板

reverse_multiscale_mapping：被映射到其他尺度的变量

var_need_init：没有被其它模型初始化或计算的变量

nodes_with_models：有模型定义的节点

type_promotion：变量类型提升

check：是否检查映射中的错误（见Details）

attribute_name：状态存储到节点中的属性名，默认为:plantsimengine_status

细节

大部分参数可以根据图和映射计算获得：

statuses通过第一次初始化获得：statuses = Dict(i => Status[] for i in nodes_with_models)

mapped_vars用mapped_variables()得到，见init_statuses中的代码

vars_need_init通过vars_need_init = Dict(org => filter(x -> isa(last(x), UninitializedVar), vars) |> keys for (org, vars) in mapped_vars) |> filter(x -> length(last(x)) > 0) 计算获得

check参数表示是否检查变量初始化。在某些变量需要初始化（部分初始化映射）的情况下，会判断节点属性（按变量名）中是否可获取。如果true，当属性不存在时会报错，否则使用模型的默认值。

PlantSimEngine.init_simulation — Method

init_simulation(mtg, mapping; nsteps=1, outputs=nothing, type_promotion=nothing, check=true, verbose=true)

初始化仿真环境，返回：

mtg

各器官类型下每个节点的status，考虑多尺度变量

模型依赖图

模型字典 organ type => NamedTuple of process => model mapping

预分配的输出outputs

参数

mtg：MTG

mapping::Dict{String,Any}：模型映射字典

nsteps：仿真步数

outputs：仿真需要的动态输出项

type_promotion：变量类型提升

check：是否检查映射错误，会传递到init_node_status!

verbose：关于映射错误的信息

细节

本函数首先为每个在mapping中拥有模型的器官类型计算status模板。 然后利用该模板初始化每个节点的status，兼顾用户自定义的初始化和多尺度mapping。 mapping用于对跨尺度定义的变量创建引用，从而保证当变量在其它尺度更新时自动同步。 目前支持两种多尺度变量类型：RefVector和MappedVar。 前者用于变量映射到一组节点，后者用于变量映射到单个节点。 用法为：只要用字符串即可表示单节点（如=> "Leaf"），用字符串数组表示多节点（如=> ["Leaf"]或=> ["Leaf", "Internode"]）。

本函数还会计算模型的依赖图，即根据模型间的输入输出关系，确定仿真调用顺序。 依赖图用于仿真运行时保证模型调用的合理顺序。

注意，如果变量既没有被模型计算，也没有在mapping初始化，会在MTG属性中查找其值。 该值是拷贝，而非对MTG属性的引用，因为无法引用Dict中的值。 如需引用特定变量，可以在MTG中直接用Ref，会被自动原样传递。

PlantSimEngine.init_statuses — Method

init_statuses(mtg, mapping, dependency_graph=dep(mapping); type_promotion=nothing, verbose=true, check=true)

获取每个MTG节点的状态，按节点类型分组，预初始化，考虑多尺度变量。

参数

mtg：植物图

mapping：模型映射的字典

dependency_graph::DependencyGraph：一级依赖图，其中每个模型在映射中被分配为一个节点。 但是，识别为硬依赖的模型不会单独作为节点存在，而是作为子节点嵌套在其他模型下。

type_promotion：用于变量的类型提升

verbose：在编译映射时打印信息

check：是否检查映射中的错误。此参数会传递给init_node_status!。

返回值

返回一个NamedTuple，内容包含按节点类型分的状态、按节点类型分的状态模板、被映射到其他尺度的变量字典、 需要被其它模型初始化或计算的变量字典，以及一个有已定义模型符号的节点向量：

(;statuses, status_templates, reverse_multiscale_mapping, vars_need_init, nodes_with_models)

PlantSimEngine.init_variables_manual — Method

init_variables_manual(models...;vars...)

以给定值初始化模型变量。

示例

using PlantSimEngine

# 加载包中的示例虚拟模型：
using PlantSimEngine.Examples

models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model()
)

PlantSimEngine.init_variables_manual(status(models), (var1=20.0,))

PlantSimEngine.initialise_all_as_hard_dependency_node — Method

initialise_all_as_hard_dependency_node(models)

针对一组模型，为每个模型初始化一个硬依赖节点， 并返回:process => HardDependencyNode组成的字典。

PlantSimEngine.is_graph_cyclic — Method

is_graph_cyclic(dependency_graph::DependencyGraph; full_stack=false, verbose=true)

判断依赖图是否包含环。

参数

dependency_graph::DependencyGraph：需要检查的依赖图。

full_stack::Bool=false：若为true，返回构成环的所有节点栈，否则只返回环本身。

warn::Bool=true：若为true，当检测到环时输出格式化的警告信息。

返回一个布尔值指示图是否有环，以及作为向量的节点栈。

PlantSimEngine.mapped_variables — Function

mapped_variables(mapping, dependency_graph=first(hard_dependencies(mapping; verbose=false)); verbose=false)

从依赖图中获取每种器官类型的变量，并为多尺度映射构建MappedVar。

参数

mapping::Dict{String,T}: 模型与尺度的映射关系。

dependency_graph::DependencyGraph: 第一阶依赖图，每个mapping中的模型为一个节点。 但被识别为“硬依赖”的模型不会分配单独节点，而是作为子节点嵌套到其他模型下。

verbose::Bool: 是否打印在mapping中寻找默认值时的堆栈跟踪。

PlantSimEngine.mapped_variables_no_outputs_from_other_scale — Function

mapped_variables_no_outputs_from_other_scale(mapping, dependency_graph=first(hard_dependencies(mapping; verbose=false)))

从依赖图中获取每种器官类型的变量，但不包括那些来自其他尺度输出的变量。

参数

mapping::Dict{String,T}: 模型与尺度的映射关系。

dependency_graph::DependencyGraph: 第一阶依赖图，每个mapping中的模型为一个节点。但硬依赖模型不会分配独立节点，而作为子节点嵌套到其他模型下。

细节

本函数返回每个器官类型的（多尺度）输入输出变量字典。

注意本函数不包括来自其他尺度但在本尺度未计算的输出变量， 相关处理见mapped_variables_with_outputs_as_inputs。

PlantSimEngine.model_ — Method

model_(m::AbstractModel)

获取抽象模型的模型（如果不是多尺度模型 MultiScaleModel，则返回其本身）。

PlantSimEngine.object_parallelizable — Method

object_parallelizable(x::T)
object_parallelizable(x::DependencyGraph)

判断模型 x 是否可以并行化计算，即模型是否能在不同对象间并行计算，返回 true 或 false。

默认情况下，所有模型均返回 false。 如果你的模型可以按对象并行化，应为其添加 ObjectDependencyTrait 方法。

本方法也可直接用于 DependencyGraph，若图中所有模型可并行化则返回 true，否则为 false。

参见

timestep_parallelizable：判断模型能否按时间步并行。

parallelizable：判断模型是否可并行化。

ObjectDependencyTrait：定义模型对其他对象的依赖性。

示例

定义一个测试过程:

using PlantSimEngine

# 定义一个测试过程:
@process "TestProcess"

定义一个对象独立的模型:

struct MyModel <: AbstractTestprocessModel end

# 重载对象依赖 trait:
PlantSimEngine.ObjectDependencyTrait(::Type{MyModel}) = IsObjectIndependent()

检查模型能否按对象并行:

object_parallelizable(MyModel()) # true

PlantSimEngine.parallelizable — Method

parallelizable(::T)
object_parallelizable(x::DependencyGraph)

判断模型 T 或整个依赖图是否可并行化，也即模型是否可在不同的时间步或对象之间并行计算。默认所有模型均返回 false。

参见

timestep_parallelizable：判断模型能否按时间步并行。

object_parallelizable：判断模型能否按对象并行。

TimeStepDependencyTrait：定义模型对其他时间步的依赖性。

示例

定义一个测试过程:

using PlantSimEngine

# 定义一个测试过程:
@process "TestProcess"

定义一个可并行化模型:

struct MyModel <: AbstractTestprocessModel end

# 重载时间步依赖 trait:
PlantSimEngine.TimeStepDependencyTrait(::Type{MyModel}) = IsTimeStepIndependent()

# 重载对象依赖 trait:
PlantSimEngine.ObjectDependencyTrait(::Type{MyModel}) = IsObjectIndependent()

检查模型是否可并行化:

parallelizable(MyModel()) # true

或者更细致地查看：

timestep_parallelizable(MyModel())
object_parallelizable(MyModel())

PlantSimEngine.ref_var — Method

ref_var(v)

为变量创建引用。如果变量已经是Base.RefValue，则直接返回，否则返回其副本的Ref， 如果是RefVector则返回其Ref。

示例

julia> using PlantSimEngine;

julia> PlantSimEngine.ref_var(1.0)
Base.RefValue{Float64}(1.0)

julia> PlantSimEngine.ref_var([1.0])
Base.RefValue{Vector{Float64}}([1.0])

julia> PlantSimEngine.ref_var(Base.RefValue(1.0))
Base.RefValue{Float64}(1.0)

julia> PlantSimEngine.ref_var(Base.RefValue([1.0]))
Base.RefValue{Vector{Float64}}([1.0])

julia> PlantSimEngine.ref_var(PlantSimEngine.RefVector([Ref(1.0), Ref(2.0), Ref(3.0)]))
Base.RefValue{PlantSimEngine.RefVector{Float64}}(RefVector{Float64}[1.0, 2.0, 3.0])

PlantSimEngine.reverse_mapping — Method

reverse_mapping(mapping::Dict{String,Tuple{Any,Vararg{Any}}}; all=true)
reverse_mapping(mapped_vars::Dict{String,Dict{Symbol,Any}})

获取模型映射字典的反向映射，即，被映射到其他尺度的变量，换句话说， 从给定尺度向其他尺度传递了哪些变量。 这可以用于，例如，了解哪些尺度需要把数值赋给其他尺度。

参数

mapping::Dict{String,Any}: 模型映射的字典。

all::Bool: 是否获取所有被映射到其他尺度的变量，包括那些被映射为单一值的变量。

返回值

一个以器官（键）为主的字典，值为字典：源器官 => 一组变量对的字典。你可以这样理解输出结果： “对于每个器官（源器官），它向哪个其他器官（目标器官）传递了自己的哪些变量。然后对于每个源器官，具体哪个变量 传递到了目标器官（对中的第一个符号），以及它在目标器官中被映射为哪个变量（对中的第二个符号）。”

示例

julia> using PlantSimEngine

导入示例模型（可在包的 examples 文件夹，或 Examples 子模块中找到）：

julia> using PlantSimEngine.Examples;

julia> mapping = Dict( "Plant" => MultiScaleModel( model=ToyCAllocationModel(), mapped_variables=[ :carbon_assimilation => ["Leaf"], :carbon_demand => ["Leaf", "Internode"], :carbon_allocation => ["Leaf", "Internode"] ], ), "Internode" => ToyCDemandModel(optimal_biomass=10.0, development_duration=200.0), "Leaf" => ( MultiScaleModel( model=ToyAssimModel(), mapped_variables=[:soil_water_content => "Soil",], ), ToyCDemandModel(optimal_biomass=10.0, development_duration=200.0), Status(aPPFD=1300.0, TT=10.0), ), "Soil" => ( ToySoilWaterModel(), ), );

注意我们在 Leaf 的 ToyAssimModel 的映射中提供了 "Soil"，而不是 ["Soil"] 。 这是因为预期在这里映射 soil_water_content 的只有一个土壤，因此允许 将该值作为单值获取，而不是一个值向量。

julia> PlantSimEngine.reverse_mapping(mapping)
Dict{String, Dict{String, Dict{Symbol, Any}}} with 3 entries:
  "Soil"      => Dict("Leaf"=>Dict(:soil_water_content=>:soil_water_content))
  "Internode" => Dict("Plant"=>Dict(:carbon_allocation=>:carbon_allocation, :ca…
  "Leaf"      => Dict("Plant"=>Dict(:carbon_allocation=>:carbon_allocation, :ca…

PlantSimEngine.save_results! — Method

save_results!(object::GraphSimulation, i)

将时间步 i 的仿真结果存储到 object 中。 对于 GraphSimulation 对象，此操作会把 status(object) 里的结果写入 outputs(object)。

PlantSimEngine.search_inputs_in_multiscale_output — Method

search_inputs_in_multiscale_output(process, organ, inputs, soft_dep_graphs)

参数

process::Symbol: 需要查找其它尺度下软依赖的进程。

organ::String: 需要查找软依赖的器官（层级）。

inputs::Dict{Symbol, Vector{Pair{Symbol}, Tuple{Symbol, Vararg{Symbol}}}}: 一个 dict，进程 => [:子进程 => (:var1, :var2)]。

soft_dep_graphs::Dict{String, ...}: 一个 dict，器官 => (依赖子图, 输入, 输出)。

rev_mapping::Dict{Symbol, Symbol}: 反向映射，映射变量 => 来源变量。

'harddependenciesfromotherscale' : 存储跨层级硬依赖的 HardDependencyNode 向量，便于不遍历整个图时访问

细节

输入（输出同理）给出了每个进程的输入，按来源进程分类，来源可以是进程自身或硬依赖的其他进程。

返回

每个进程在其他层级下在输出中找到的软依赖变量组成的字典，例如：

Dict{String, Dict{Symbol, Vector{Symbol}}} with 2 entries:
    "Internode" => Dict(:carbon_demand=>[:carbon_demand])
    "Leaf"      => Dict(:carbon_assimilation=>[:carbon_assimilation], :carbon_demand=>[:carbon_demand])

上述结果表示变量 :carbon_demand 由 "Internode" 层级下的 :carbondemand 过程计算，变量 `:carbonassimilation` 则由 "Leaf" 层级下的 :carbon_assimilation 过程计算。这些变量作为当前进程的输入。

PlantSimEngine.search_inputs_in_output — Method

search_inputs_in_output(process, inputs, outputs)

返回依赖图 d 中进程的软依赖组成的字典。 软依赖是指模型中没有显式定义，但可以通过进程的输入输出推断出来的依赖关系。

参数

process::Symbol: 需要查找软依赖的进程。

inputs::Dict{Symbol, Vector{Pair{Symbol}, Tuple{Symbol, Vararg{Symbol}}}}: 一个 dict，进程对应其输入符号。

outputs::Dict{Symbol, Tuple{Symbol, Vararg{Symbol}}}: 一个 dict，进程对应其输出符号。

细节

输入（输出同理）给出了每个进程的输入，按来源进程分类，来源可以是进程自身或硬依赖的其他进程。

返回

进程软依赖组成的字典。

示例

in_ = Dict(
    :process3 => [:process3=>(:var4, :var5), :process2=>(:var1, :var3), :process1=>(:var1, :var2)],
    :process4 => [:process4=>(:var0,)],
    :process6 => [:process6=>(:var7, :var9)],
    :process5 => [:process5=>(:var5, :var6)],
)

out_ = Dict(
    :process3 => Pair{Symbol}[:process3=>(:var4, :var6), :process2=>(:var4, :var5), :process1=>(:var3,)],
    :process4 => [:process4=>(:var1, :var2)],
    :process6 => [:process6=>(:var8,)],
    :process5 => [:process5=>(:var7,)],
)

search_inputs_in_output(:process3, in_, out_)
(process4 = (:var1, :var2),)

PlantSimEngine.set_variables_at_timestep! — Method

set_variables_at_timestep!(status_timestep::Status, user_status::Status, variables_to_update, timestep)

将 status_timestep 中所有在 variables_to_update 中的变量，更新为用户提供的 user_status 在指定 timestep 上的当前值。 变量名列表由 variables_to_update 提供，为符号向量。

status_timestep 是表示单次时间步长的状态，user_status 是用户提供的状态，包含未被任何模型计算的变量的值。 其变量值可以为常量，也可以是向量。如果为向量，则用 timestep 指定当前步长所用的值。

PlantSimEngine.soft_dependencies — Function

soft_dependencies(d::DependencyGraph)

返回一个带有依赖图 d 中进程的软依赖的 DependencyGraph。 软依赖是指模型中没有显式定义，但可以通过进程的输入输出推断出来的依赖关系。

参数

d::DependencyGraph: 硬依赖图。

示例

using PlantSimEngine

# 加载包中提供的示例模型：
using PlantSimEngine.Examples

# 创建模型列表：
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    process4=Process4Model(),
    process5=Process5Model(),
    process6=Process6Model(),
)

# 创建硬依赖图：
hard_dep = hard_dependencies(models.models, verbose=true)

# 获取软依赖图：
soft_dep = soft_dependencies(hard_dep)

PlantSimEngine.status_from_template — Method

status_from_template(d::Dict{Symbol,Any})

从变量和值的模板字典创建一个status。如果值本身是RefValue或RefVector，则直接使用，否则会自动转为Ref。

参数

d::Dict{Symbol,Any}：变量和值的字典。

返回

一个Status。

示例

julia> using PlantSimEngine

julia> a, b = PlantSimEngine.status_from_template(Dict(:a => 1.0, :b => 2.0));

julia> a
1.0

julia> b
2.0

PlantSimEngine.timestep_parallelizable — Method

timestep_parallelizable(x::T)
timestep_parallelizable(x::DependencyGraph)

返回 true 表示模型 x 可以按时间步进行并行计算，否则返回 false。

默认情况下，所有模型都返回 false。 如果您开发的模型能够按时间步并行，请为您的模型添加 ObjectDependencyTrait 的方法。

本方法也可直接用于 DependencyGraph，若图中所有模型可并行化则返回 true，否则为 false。

参见

object_parallelizable：判断模型能否按时间步并行。

parallelizable：判断模型是否可并行化。

TimeStepDependencyTrait：定义模型对其他时间步的依赖性。

示例

定义一个测试过程:

using PlantSimEngine

# 定义一个测试过程:
@process "TestProcess"

定义一个时间步独立的模型:

struct MyModel <: AbstractTestprocessModel end

# 重载时间步依赖 trait:
PlantSimEngine.TimeStepDependencyTrait(::Type{MyModel}) = IsTimeStepIndependent()

检查模型能否按对象并行:

timestep_parallelizable(MyModel()) # true

PlantSimEngine.transform_single_node_mapped_variables_as_self_node_output! — Method

transform_single_node_mapped_variables_as_self_node_output!(mapped_vars)

查找作为SingleNodeMapping输入到其他尺度的变量，并以自身为源在源尺度声明为MappedVar。 这样有助于我们在创建模板status对象时以引用方式声明。

这些节点表现为[:variable_name => "Plant"]的写法（注意"Plant"为标量）。

PlantSimEngine.traverse_dependency_graph! — Method

traverse_dependency_graph(node::HardDependencyNode, f::Function, var::Vector)

对 node 应用函数 f，然后遍历其子节点（硬依赖节点）。

通过将节点的进程名和函数 f 的结果组成的对推入向量 var，以实现对 var 的累加。

PlantSimEngine.traverse_dependency_graph! — Method

traverse_dependency_graph(node::SoftDependencyNode, f::Function, var::Vector; visit_hard_dep=true)

对 node 应用函数 f，遍历其硬依赖节点（当 visit_hard_dep=true 时），然后遍历其软依赖的子节点。

通过将节点的进程名和函数 f 的结果组成的对推入向量 var，以实现对 var 的累加。

PlantSimEngine.traverse_dependency_graph — Method

traverse_dependency_graph(graph::DependencyGraph, f::Function; visit_hard_dep=true)
traverse_dependency_graph(graph; visit_hard_dep=true)

遍历依赖 graph，并对每个节点应用函数 f。如果未提供 f，则仅返回节点本身。

首先遍历第一层软依赖节点，然后遍历它们的硬依赖节点（如果 visit_hard_dep=true），最后遍历软依赖节点的子节点。

返回一个由节点与函数 f 结果组成的元组对的向量。

示例

using PlantSimEngine

# 包含示例进程和模型:
using PlantSimEngine.Examples;

function f(node)
    node.value
end

vars = (
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    process4=Process4Model(),
    process5=Process5Model(),
    process6=Process6Model(),
    process7=Process7Model(),
)

graph = dep(vars)
traverse_dependency_graph(graph, f)

PlantSimEngine.variables_multiscale — Function

variables_multiscale(node, organ, mapping, st=NamedTuple())

获取 HardDependencyNode 的变量，考虑多尺度映射，也就是说， 如果变量被映射到了另一个尺度，则定义为 MappedVar。默认值 取自模型本身，如果用户未指定（st），且如果为节点的输入变量，则 用 UninitializedVar 标记。

返回一个包含变量及其默认值的 NamedTuple。

参数

node::HardDependencyNode：需要获取变量的节点。

organ::String：器官类型，如 "Leaf"。

vars_mapping::Dict{String,T}：模型的映射（详见下方）。

st::NamedTuple：可选，包含变量默认值的 NamedTuple。

详情

vars_mapping 是一个以器官类型为 key、以字典为 value 的字典。 它由用户映射计算得到：

PlantSimEngine.variables_outputs_from_other_scale — Method

variables_outputs_from_other_scale(mapped_vars)

对于mapped_vars中的每个器官，查找那些仅为其他尺度输出、而本尺度未计算的变量。 该函数用于mapped_variables。

PlantSimEngine.variables_typed — Method

variables_typed(model)
variables_typed(model, models...)

返回模型所需变量（变量名及类型）的具名元组，或多个模型的变量类型联合。

示例

using PlantSimEngine;

# Load the dummy models given as example in the package:
using PlantSimEngine.Examples;

PlantSimEngine.variables_typed(Process1Model(1.0))
(var1 = Float64, var2 = Float64, var3 = Float64)

PlantSimEngine.variables_typed(Process1Model(1.0), Process2Model())

# output
(var4 = Float64, var5 = Float64, var1 = Float64, var2 = Float64, var3 = Float64)

参见

inputs, outputs and variables

PlantSimEngine.vars_not_init_ — Method

vars_not_init_(st<:Status, var_names)

获取状态结构体中尚未正确初始化的变量。