公共 API

索引

PlantMeteo.TimeStepTable

PlantSimEngine.AbstractModel

PlantSimEngine.ModelList

PlantSimEngine.MultiScaleModel

PlantSimEngine.PreviousTimeStep

PlantSimEngine.Status

PlantSimEngine.EF

PlantSimEngine.NRMSE

PlantSimEngine.RMSE

PlantSimEngine.add_organ!

PlantSimEngine.dep

PlantSimEngine.dr

PlantSimEngine.fit

PlantSimEngine.init_status!

PlantSimEngine.init_variables

PlantSimEngine.inputs

PlantSimEngine.inputs

PlantSimEngine.is_initialized

PlantSimEngine.outputs

PlantSimEngine.outputs

PlantSimEngine.run!

PlantSimEngine.status

PlantSimEngine.to_initialize

PlantSimEngine.variables

PlantSimEngine.variables

PlantSimEngine.variables

PlantSimEngine.@process

API 文档

PlantMeteo.TimeStepTable — Method

TimeStepTable{Status}(df::DataFrame)

方法：从 DataFrame 构建 TimeStepTable（来自 PlantMeteo.jl），但每一行都是 Status。

注意

ModelList 默认使用 TimeStepTable{Status}（见下面的示例）。

示例

using PlantSimEngine, DataFrames

# 从 DataFrame 创建一个 TimeStepTable：
df = DataFrame(
    Tₗ=[25.0, 26.0],
    aPPFD=[1000.0, 1200.0],
    Cₛ=[400.0, 400.0],
    Dₗ=[1.0, 1.2],
)
TimeStepTable{Status}(df)

# 只要其中一个变量有多个值，叶片会自动用带时间步的 TimeStepTable{Status}：
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

# 叶片的 status 是一个 TimeStepTable：
status(models)

# 当然也可以手动用 Status 创建 TimeStepTable：
TimeStepTable(
    [
        Status(Tₗ=25.0, aPPFD=1000.0, Cₛ=400.0, Dₗ=1.0),
        Status(Tₗ=26.0, aPPFD=1200.0, Cₛ=400.0, Dₗ=1.2),
    ]
)

PlantSimEngine.AbstractModel — Type

抽象模型类型。所有模型都是此类型的子类型。

PlantSimEngine.ModelList — Type

ModelList(models::M, status::S)
ModelList(;
    status=nothing,
    type_promotion=nothing,
    variables_check=true,
    kwargs...
)

列举用于模拟的一组模型 (models)，并自动完成变量初始化、类型提升、时间步长处理等样板工作。

Note

status 字段取决于输入的模型。你可以通过已实例化模型上的 variables 方法获取模型所需变量，也可以使用 inputs 和 outputs。

参数说明

models: 一个模型的列表。通常采用 NamedTuple，但也可以是实现了 getproperty 的其他结构。

status: 包含模型变量初始化的结构体。作为关键字参数时通常为 NamedTuple，或为实现了 Tables.jl 接口的任何结构体（如 DataFrame，详见下文）。

type_promotion: 可选，用于默认值变量的类型转换。默认为 nothing（即不转换）。注意，用户以 kwargs 提供的变量不会自动转换（需要手动转换）。需提供以当前类型为键、新类型为值的 Dict。

variables_check=true: 检查用户是否初始化了所有必需变量。

kwargs: 各过程名称对应的模型。

细节

如果你需要自定义 status 的类型，并希望用户能够只初始化部分字段，则必须为 add_model_vars! 实现一个方法，以在类型未完全初始化时添加模型变量。默认方法兼容任何实现了 Tables.jl 接口的类型（如 DataFrame），以及 NamedTuples。

注意，如果输入 status 未涵盖所有变量，ModelList 会对其进行复制。

示例

以下示例采用包中 examples/dummy.jl 的示例模型。其实现了三个虚拟过程：Process1Model、Process2Model 和 Process3Model，且每个过程有一个模型实现。

julia> using PlantSimEngine;

包含示例过程和模型：

julia> using PlantSimEngine.Examples;

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model());
[ Info: Some variables must be initialized before simulation: (process1 = (:var1, :var2), process2 = (:var1,)) (see `to_initialize()`)

julia> typeof(models)
ModelList{@NamedTuple{process1::Process1Model, process2::Process2Model, process3::Process3Model}, Status{(:var5, :var4, :var6, :var1, :var3, :var2), NTuple{6, Base.RefValue{Float64}}}}

未以关键字参数提供变量，意味着 ModelList 的 status 尚未设置，所有变量将按 inputs 和 outputs 中的默认值初始化（通常是 typemin(Type)，比如浮点数则为 -Inf）。此时组件尚不可模拟。

模拟前需初始化哪些变量，可通过 to_initialize 方法查询：

julia> to_initialize(models)
(process1 = (:var1, :var2), process2 = (:var1,))

我们现在可以在 status 字段中为这些变量赋值，并对 ModelList 进行模拟。例如针对 process3（与 process1 和 process2 耦合）：

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=(var1=15.0, var2=0.3));

julia> meteo = Atmosphere(T = 22.0, Wind = 0.8333, P = 101.325, Rh = 0.4490995);

julia> outputs_sim = run!(models,meteo)
TimeStepTable{Status{(:var5, :var4, :var6, ...}(1 x 6):
╭─────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────╮
│ Row │    var5 │    var4 │    var6 │    var1 │    var3 │    var2 │
│     │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │ Float64 │
├─────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│   1 │ 36.0139 │    22.0 │ 58.0139 │    15.0 │     5.5 │     0.3 │
╰─────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────╯

julia> outputs_sim[:var6]
1-element Vector{Float64}:
 58.0138985

如需对变量使用特殊类型，可使用 type_promotion 参数：

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=(var1=15.0, var2=0.3), type_promotion = Dict(Float64 => Float32));

我们使用 type_promotion 将 status 强制转换为 Float32：

julia> [typeof(models[i][1]) for i in keys(status(models))]
6-element Vector{DataType}:
 Float32
 Float32
 Float32
 Float64
 Float64
 Float32

可以看到，只有默认变量（未在 status 参数中给定的变量）被转换为 Float32，其余两个用户提供的变量未被转换。这样做是为了让用户可以为 status 中赋值的变量指定任意类型。如需全部变量都转为 Float32，可直接以 Float32 赋值：

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=(var1=15.0f0, var2=0.3f0), type_promotion = Dict(Float64 => Float32));

我们使用 type_promotion 将 status 强制转换为 Float32：

julia> [typeof(models[i][1]) for i in keys(status(models))]
6-element Vector{DataType}:
 Float32
 Float32
 Float32
 Float32
 Float32
 Float32

PlantSimEngine.MultiScaleModel — Type

MultiScaleModel(model, mapped_variables)

一个用于实现多尺度模型的结构体。它定义了模型变量与节点标记之间的映射，变量值将从这些节点获取。

参数

model<:AbstractModel：待多尺度化的模型

mapped_variables<:Vector{Pair{Symbol,Union{AbstractString,Vector{AbstractString}}}}：由符号和字符串（或字符串向量）组成的配对向量

mapped_variables 参数可以有以下形式：

[:variable_name => "Plant"]：从 Plant 节点获取单个值

[:variable_name => ["Leaf"]]：从所有 Leaf 节点获取一个值向量

[:variable_name => ["Leaf", "Internode"]]：从 Leaf 和 Internode 节点获取一个值向量

[:variable_name => "Plant" => :variable_name_in_plant_scale]：从 Plant 节点以指定变量名获取值

[:variable_name => ["Leaf" => :variable_name_1, "Internode" => :variable_name_2]]：从 Leaf 和 Internode 节点分别以不同变量名获取值，返回一个向量

[PreviousTimeStep(:variable_name) => ...]：标记变量用前一时间步的值初始化，该变量不参与依赖图的构建

[:variable_name => :variable_name_from_another_model]：从相同尺度的其它模型变量获取数据，并重命名

[PreviousTimeStep(:variable_name),]：纯粹标记为 PreviousTimeStep，不参与依赖图构建

不同形式的详细说明：

模型的 variable_name 变量将从 Plant 节点获取，假设只有一个节点是 Plant。此时 status 里该值是标量，因此用户需确保 MTG 中只有一个 Plant 类型节点。

模型的 variable_name 变量将从所有 Leaf 节点获取。注意此处为向量，模型需能处理值向量。即使只有一个 Leaf 节点，取到的依然是只有一个元素的向量。

模型的 variable_name 变量将从所有 Leaf 和 Internode 节点获取。即取所有这两类节点的值向量。

模型的 variable_name 变量将从 Plant 节点中的 variable_name_in_plant_scale 变量获取。当模型变量名与节点变量名不同时可用。

模型的 variable_name 变量将分别从 Leaf 节点的 variable_name_1 及 Internode 节点的 variable_name_2 取得。

模型的 variable_name 变量使用上一步计算结果，不用于构建当前步依赖图。可用于变量依赖自身时避免循环依赖。需要时可在 Status 初始化其值。

模型的 variable_name 变量从同一尺度但不同变量名的其它模型获取。

模型的 variable_name 变量仅作 PreviousTimeStep 标记，不参与依赖图的构建。

请注意，该映射不会复制变量值，仅引用。当某一节点的 status 被更新，其他节点引用的值也随之改变。

示例

julia> using PlantSimEngine;

包含示例过程和模型：

julia> using PlantSimEngine.Examples;

取一个模型示例：

julia> model = ToyCAllocationModel()
ToyCAllocationModel()

我们通过定义模型变量和节点标记之间的映射，将其转为多尺度模型。

例如，假设 carbon_allocation 来源于 Leaf 和 Internode 节点，可以定义映射如下：

julia> mapped_variables=[:carbon_allocation => ["Leaf", "Internode"]]
1-element Vector{Pair{Symbol, Vector{String}}}:
 :carbon_allocation => ["Leaf", "Internode"]

mappedvariables 参数是符号和字符串（或字符串向量）的配对向量。以上例只有一对，表示将 `carbonallocation变量与Leaf和Internode` 关联。

现在将模型及变量映射传递给 MultiScaleModel 构造函数，实现多尺度模型：

julia> multiscale_model = PlantSimEngine.MultiScaleModel(model, mapped_variables)
MultiScaleModel{ToyCAllocationModel, Vector{Pair{Union{Symbol, PreviousTimeStep}, Union{Pair{String, Symbol}, Vector{Pair{String, Symbol}}}}}}(ToyCAllocationModel(), Pair{Union{Symbol, PreviousTimeStep}, Union{Pair{String, Symbol}, Vector{Pair{String, Symbol}}}}[:carbon_allocation => ["Leaf" => :carbon_allocation, "Internode" => :carbon_allocation]])

可访问映射变量和模型：

julia> PlantSimEngine.mapped_variables_(multiscale_model)
1-element Vector{Pair{Union{Symbol, PreviousTimeStep}, Union{Pair{String, Symbol}, Vector{Pair{String, Symbol}}}}}:
 :carbon_allocation => ["Leaf" => :carbon_allocation, "Internode" => :carbon_allocation]

julia> PlantSimEngine.model_(multiscale_model)
ToyCAllocationModel()

PlantSimEngine.PreviousTimeStep — Type

PreviousTimeStep(variable)

用于手动标记模型中某变量以采用上一个时间步中计算的值的结构体。 这意味着该变量不会被用于构建依赖关系图，因为依赖关系图仅适用于当前时间步。 当变量依赖于自身时，为了避免循环依赖，可以使用该方法。 如有需要，其值可以在 Status 中初始化。

当构建 MultiScaleModel 时会添加该过程，以避免不同进程之间具有相同变量名时的冲突。 例如，一个进程可以将变量 :carbon_biomass 定义为 PreviousTimeStep，而另一个进程则将该变量作为当前时间步的依赖项使用（这样是允许的，因为它们不会出现循环依赖的问题）。

PlantSimEngine.Status — Type

Status(vars)

Status类型用于在模拟过程中存储变量的值。主要用作TimeStepTable的TimeStepRow中用于保存变量的结构体（参见PlantMeteo.jl 文档），也用于ModelList。

大部分代码来自 MasonProtter/MutableNamedTuples.jl，因此Status本质上是一个经过少量修改的MutableNamedTuples， 也就是一个用来存储对变量值的引用的NamedTuple结构体，使其为可变类型。

示例

所有变量只有一个值的叶片将得到一个只有一个时间步长的状态对象：

julia> using PlantSimEngine

julia> st = PlantSimEngine.Status(Ra_SW_f=13.747, sky_fraction=1.0, d=0.03, aPPFD=1500.0);

以下索引方法均可用：

julia> st[:Ra_SW_f]
13.747

julia> st.Ra_SW_f
13.747

julia> st[1]
13.747

设置 Status 变量也很简单：

julia> st[:Ra_SW_f] = 20.0
20.0

julia> st.Ra_SW_f = 21.0
21.0

julia> st[1] = 22.0
22.0

PlantSimEngine.EF — Method

EF(obs,sim)

使用NSE（Nash-Sutcliffe效率）模型，返回观测值 obs 和模拟值 sim 之间的效率系数（Efficiency Factor）。 更多信息见 https://en.wikipedia.org/wiki/Nash%E2%80%93Sutcliffemodelefficiency_coefficient 。

值越接近1越好。

示例

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

EF(obs, sim)

PlantSimEngine.NRMSE — Method

NRMSE(obs,sim)

返回观测值 obs 和模拟值 sim 之间的归一化均方根误差（Normalized Root Mean Squared Error）。 归一化方法是除以观测值的范围（最大值-最小值）。

示例

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

NRMSE(obs, sim)

PlantSimEngine.RMSE — Method

RMSE(obs,sim)

返回观测值 obs 和模拟值 sim 之间的均方根误差（Root Mean Squared Error）。

值越接近0越好。

示例

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

RMSE(obs, sim)

PlantSimEngine.add_organ! — Method

add_organ!(node::MultiScaleTreeGraph.Node, sim_object, link, symbol, scale; index=0, id=MultiScaleTreeGraph.new_id(MultiScaleTreeGraph.get_root(node)), attributes=Dict{Symbol,Any}(), check=true)

向图中添加一个器官，并自动初始化该器官的（多尺度）变量状态。

此函数应从实现器官出现的模型中调用，例如基于热时间的函数。

参数说明

node：要向其添加器官的节点（新器官的父节点）

sim_object：模拟对象，例如模型的extra参数中的 GraphSimulation 对象。

link：新节点与器官之间的连接类型：

"<"：新节点在父器官之后

"+"：新节点从父器官分枝

"/"：新节点分解父器官，即更改尺度

symbol：器官的符号，例如 "Leaf"

scale：器官的尺度，例如 2。

index：器官的序号，例如 1。index 可用于方便地标识分枝顺序，或轴上的生长单元序号。它不同于唯一的节点 id。

id：新节点的唯一编号。如果未提供，则自动生成新编号。

attributes：新节点的属性。如果未提供，则使用空字典。

check：布尔值，表示是否检查变量初始化。会传递给 init_node_status!。

返回值

status：新节点的状态

示例

具体用法请参考 Examples 模块中的 ToyInternodeEmergence 示例模型（也可在 examples 文件夹找到），或 test-mtg-dynamic.jl 测试文件。

PlantSimEngine.dep — Function

dep(m::ModelList)
dep(mapping::Dict{String,T}; verbose=true)
dep!(m::ModelList, nsteps=1)

给定 ModelList 或多尺度模型映射，获取模型的依赖图。如果返回一个图，则所有模型耦合在一起；如果返回多个图，则每个图内的模型相互耦合，不同图中的模型不耦合。 nsteps 是依赖图将被使用的步数，用于决定每个软依赖在图中的 simulation_id 参数长度。对于多尺度映射，该值为 1。

细节

依赖图通过在每个过程的输入中查找其本层或其他层输出构建。针对每个模型（一个模型模拟一个过程）有五种情况：

过程无输入。它是完全独立的，作为依赖图的根节点之一。

过程需要同层模型输入。被放置为另一个过程的子节点。

过程需要来自另一层的输入。被设置为另一个层过程的子节点。

过程需要本层和另一层的输入。分别作为两个过程的子节点。

过程作为另一个过程的硬依赖（仅可能在同层）。此时该过程被设为另一个过程的硬依赖，仿真由该过程直接控制。

第四种情况，过程有两个父过程。这样处理没有问题，因为仿真时会检查两个父节点是否都已运行，仅当均已运行后再运行当前过程。

第五种情况，依然需要检查变量是否来自另一层。此时父节点作为另一层过程的子节点。需注意，可能存在多层硬依赖图，因此此过程是递归完成的。

如何实现以上功能？首先识别硬依赖，然后将硬依赖根节点的输入输出与其它层连接（如有必要）。随后将所有这些节点转为软依赖，放入一个 Dict（Scale => Dict(process => SoftDependencyNode)）中。 遍历所有节点，将需要其它节点输出作为输入的节点，设置为子/父节点关系。 如某节点无任何依赖，则设置为根节点并推入新 Dict（independantprocessroot）。此 Dict 就是返回的依赖图，根节点为各子图的独立起点，这些子图即为被耦合在一起的模型。此后可分别遍历每个子图进行仿真。

备注

dep(m::ModelList) 与 dep!(m::ModelList, nsteps) 的区别在于，前者返回模型列表中的依赖图，后者返回指定步数的依赖图，并对每个节点的 simulationid 进行修正（`simulationid=fill(0, nsteps)`）。

示例

using PlantSimEngine

# 包含示例过程与模型：
using PlantSimEngine.Examples;

models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

dep(models)

# 或直接用过程：
models = (
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    process4=Process4Model(),
    process5=Process5Model(),
    process6=Process6Model(),
    process7=Process7Model(),
)

dep(;models...)

PlantSimEngine.dr — Method

dr(obs,sim)

返回Willmott提出的改进一致性指数dᵣ。 参考：Willmot et al. 2011. A refined index of model performance. https://rmets.onlinelibrary.wiley.com/doi/10.1002/joc.2419

值越接近1越好。

示例

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

dr(obs, sim)

PlantSimEngine.fit — Function

fit()

利用观测值（以及可能的初始化值）优化模型参数。

建模者应为其模型实现一个 fit 方法，建议采用以下设计模式：

函数调用应将模型类型作为第一个参数（T::Type{<:AbstractModel}）， 数据作为第二个参数（支持 Table.jl 的类型，如 DataFrame），并将参数初始化作为关键字参数传递（必要时可设默认值）。

例如，拟合示例脚本中的 Beer 模型的方法（参考 src/examples/Beer.jl）如下：

function PlantSimEngine.fit(::Type{Beer}, df; J_to_umol=PlantMeteo.Constants().J_to_umol)
    k = Statistics.mean(log.(df.Ri_PAR_f ./ (df.aPPFD ./ J_to_umol)) ./ df.LAI)
    return (k=k,)
end

该函数应返回以 NamedTuple 形式包含优化参数的元组，如 (parameter_name=parameter_value,)。

以下是使用 Beer 模型的一个示例，其中通过"aPPFD"、"LAI" 和 "RiPARf" 的观测值拟合参数 k。

# 引入示例过程与模型：
using PlantSimEngine.Examples;

m = ModelList(Beer(0.6), status=(LAI=2.0,))
meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_f=300.0)
run!(m, meteo)
df = DataFrame(aPPFD=m[:aPPFD][1], LAI=m.status.LAI[1], Ri_PAR_f=meteo.Ri_PAR_f[1])
fit(Beer, df)

注意，这是一个用于演示拟合方法有效性的虚拟示例。在该例中，通过设置 k=0.6 用 Beer-Lambert 定律模拟 aPPFD，并再次利用模拟得到的 aPPFD 来拟合 k，最终得到与模拟时一致的参数值。

PlantSimEngine.init_status! — Method

init_status!(object::Dict{String,ModelList};vars...)
init_status!(component::ModelList;vars...)

为组件赋予用户输入的初值。

示例

using PlantSimEngine

# 加载包中的示例虚拟模型：
using PlantSimEngine.Examples

models = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    ),
    "InterNode" => ModelList(
        process1=Process1Model(1.0),
    )
)

init_status!(models, var1=1.0 , var2=2.0)
status(models["Leaf"])

PlantSimEngine.init_variables — Method

init_variables(models...)

用模型的默认值初始化变量。这些变量取自模型的输入和输出。

示例

using PlantSimEngine

# 加载包中的示例虚拟模型：
using PlantSimEngine.Examples

init_variables(Process1Model(2.0))
init_variables(process1=Process1Model(2.0), process2=Process2Model())

PlantSimEngine.inputs — Method

inputs(model::AbstractModel)
inputs(...)

获取一个或多个模型的输入变量。

对于 AbstractModel（无输入变量）或 Missing 类型的模型，默认返回空元组。

示例

using PlantSimEngine;

# Load the dummy models given as example in the package:
using PlantSimEngine.Examples;

inputs(Process1Model(1.0))

# output
(:var1, :var2)

PlantSimEngine.inputs — Method

inputs(mapping::Dict{String,T})

获取映射中每个过程和器官类型的模型输入变量。

PlantSimEngine.is_initialized — Method

is_initialized(m::T) where T <: ModelList
is_initialized(m::T, models...) where T <: ModelList

检查必须初始化的变量是否已经被初始化。如果已初始化全部变量，返回 true，否则返回 false 并给出信息提示。

注意

无法预先知道用户将仿真哪些过程，因此，如果一个组件为每个过程都定义了模型，则需要初始化的变量始终是所有变量中最小的那个子集，即认为用户将仿真其它模型所需的变量。

示例

using PlantSimEngine

# 加载包中的示例虚拟模型：
using PlantSimEngine.Examples

models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model()
)

is_initialized(models)

PlantSimEngine.outputs — Method

outputs(model::AbstractModel)
outputs(...)

获取一个或多个模型的输出变量。

对于 AbstractModel（无输出变量）或 Missing 类型的模型，默认返回空元组。

示例

using PlantSimEngine;

# Load the dummy models given as example in the package:
using PlantSimEngine.Examples;

outputs(Process1Model(1.0))

# output
(:var3,)

PlantSimEngine.outputs — Method

outputs(mapping::Dict{String,T})

获取映射中每个过程和器官类型的模型输出变量。

PlantSimEngine.run! — Function

run!(object, meteo, constants, extra=nothing; check=true, executor=Floops.ThreadedEx())
run!(object, mapping, meteo, constants, extra; nsteps, outputs, check, executor)

运行仿真，对模型列表中的每个模型按照正确顺序（即按照依赖关系图）执行。

如果给定多个时间步，模型将在每个时间步顺序运行。

参数说明

object：一个 ModelList、ModelList 的数组或字典，或者一个植物图（MTG）。

meteo：一个 PlantMeteo.TimeStepTable，其元素为 PlantMeteo.Atmosphere，或单一 PlantMeteo.Atmosphere。

constants：一个 PlantMeteo.Constants 对象，或由常量键和值组成的 NamedTuple。

extra：额外参数。在仿真植物图（MTG）时不可用（仿真对象通过此参数传递）。

check：若为 true，在运行仿真前检查模型列表合法性（会花费少量时间），且在运行过程中返回更多信息。

executor：用于运行仿真的 Floops 执行器，可以顺序执行（executor=SequentialEx()）、多线程并行（executor=ThreadedEx()，默认）、分布式并行（executor=DistributedEx()）。

mapping：MTG 与模型列表之间的映射关系。

nsteps：需要运行的时间步数，仅在未给定 meteo 时需要（否则会自动从 meteo 推断）。

outputs：对于 MTG 的每个节点类型，需动态获得的输出变量。

返回值

在原地修改对象的 status。用户可通过 status 函数（参见示例）从对象获取结果。

细节

模型执行

模型按照依赖图顺序运行。如果一个模型对另一个模型有软依赖（即输入由另一模型计算），会优先运行被依赖模型。若有多个软依赖，则会优先计算所有父模型（软依赖）。

并行执行

用户可以通过为 executor 参数提供兼容执行器实现并行。软件包会自动检查是否允许并行。如果不允许而用户指定了并行，将会发出警告，并转为顺序执行。 我们使用 Floops 包实现并行仿真，意味着你可将任何相容的执行器传入 executor 参数。 可参考 FoldsThreads.jl（线程执行器）、FoldsDagger.jl（Dagger 框架并行 fold）、以及即将发布的 FoldsCUDA.jl（GPU 计算，详见 本议题）和 FoldsKernelAbstractions.jl。也可以通过 ParallelMagics.jl 检查是否可自动并行。

示例

导入相关包：

julia> using PlantSimEngine, PlantMeteo;

加载 Examples 子模块中给出的示例模型：

julia> using PlantSimEngine.Examples;

创建模型列表：

julia> models = ModelList(Process1Model(1.0), Process2Model(), Process3Model(), status = (var1=1.0, var2=2.0));

创建气象数据：

julia> meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_f=300.0);

运行仿真：

julia> outputs_sim = run!(models, meteo);

获取仿真输出：

julia> (outputs_sim[:var4],outputs_sim[:var6])
([12.0], [41.95])

PlantSimEngine.status — Method

status(m)
status(m::AbstractArray{<:ModelList})
status(m::AbstractDict{T,<:ModelList})

获取 ModelList 的状态，即输入（和输出）变量的状态。

另见 is_initialized 和 to_initialize

示例

using PlantSimEngine

# Including example models and processes:
using PlantSimEngine.Examples;

# Create a ModelList
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status = (var1=[15.0, 16.0], var2=0.3)
);

status(models)

# Or just one variable:
status(models,:var1)

# Or the status at the ith time-step:
status(models, 2)

# Or even more simply:
models[:var1]
# output
2-element Vector{Float64}:
 15.0
 16.0

PlantSimEngine.to_initialize — Method

to_initialize(; verbose=true, vars...)
to_initialize(m::T)  where T <: ModelList
to_initialize(m::DependencyGraph)
to_initialize(mapping::Dict{String,T}, graph=nothing)

返回必须初始化的变量，提供一组模型与过程。 该函数考虑模型耦合，仅返回必需的变量，即只返回那些由于部分模型的输出变量被其他模型作为输入变量而需要的变量。

参数

verbose: 如果为true，打印信息消息。

vars...: 要考虑的模型和过程。

m::T: 一个ModelList。

m::DependencyGraph: 一个 DependencyGraph。

mapping::Dict{String,T}: 一个模型与器官的映射。

graph: 表示植物或场景的图（如多尺度树图）。该图被用于检测未初始化变量是否在图节点属性中已存在。

示例

using PlantSimEngine

# 加载包中的示例虚拟模型：
using PlantSimEngine.Examples

to_initialize(process1=Process1Model(1.0), process2=Process2Model())

# 或直接使用组件：
models = ModelList(process1=Process1Model(1.0), process2=Process2Model())
to_initialize(models)

m = ModelList(
    (
        process1=Process1Model(1.0),
        process2=Process2Model()
    ),
    Status(var1 = 5.0, var2 = -Inf, var3 = -Inf, var4 = -Inf, var5 = -Inf)
)

to_initialize(m)

或者带有映射：

using PlantSimEngine

# 加载包中的示例虚拟模型：
using PlantSimEngine.Examples

mapping = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    ),
    "Internode" => ModelList(
        process1=Process1Model(1.0),
    )
)

to_initialize(mapping)

PlantSimEngine.variables — Method

variables(pkg::Module)

返回某个依赖 PlantSimEngine 的包中所有变量、变量描述及单位的数据表（需包作者实现）。

开发者注意事项

依赖 PlantSimEngine 的包开发者应将变量信息写入 "data/variables.csv" 文件， 该函数会返回此文件的内容。

示例

以下为 PlantBiophysics 包的示例：

#] add PlantBiophysics
using PlantBiophysics
variables(PlantBiophysics)

PlantSimEngine.variables — Method

variables(mapping::Dict{String,T})

获取映射中每个过程和器官类型模型的变量（输入与输出）。

PlantSimEngine.variables — Method

variables(model)
variables(model, models...)

返回模型所需变量的名称元组，或多个模型变量名的并集。

注意

每个模型都可以（且应该）为此函数定义一个方法。

using PlantSimEngine;

# Load the dummy models given as example in the package:
using PlantSimEngine.Examples;

variables(Process1Model(1.0))

variables(Process1Model(1.0), Process2Model())

# output

(var1 = -Inf, var2 = -Inf, var3 = -Inf, var4 = -Inf, var5 = -Inf)

参见

inputs, outputs and variables_typed

PlantSimEngine.@process — Macro

@process(process::String, doc::String=""; verbose::Bool=true)

该宏用于生成某一过程的抽象类型及部分仿真的样板代码，并自动生成相关文档。如果verbose=true，还会输出简短的模型实现教程。

该抽象过程类型会作为所有该过程模型实现的父类型，其名称形式为 "Abstract<ProcessName>Model"，例如对于过程名 "growth"，其父类型为 AbstractGrowthModel。

@process 的第一个参数为新过程名称，第二个参数为附加到 Abstract<ProcessName>Model 类型上的额外文档内容，第三个参数决定是否打印简短教程。

建议新用户使用该宏，因为它会详细说明后续的操作流程。更有经验的用户可直接定义 abstract type 并指定为 AbstractModel 的子类，而不输出教程：

abstract type MyNewProcess <: AbstractModel end

例子

@process "dummy_process" "This is a dummy process that shall not be used"