API

Index

DataFrames.DataFrame

DataFrames.DataFrame

DataFrames.DataFrame

PlantMeteo.TimeStepTable

PlantSimEngine.AbstractModel

PlantSimEngine.DependencyTree

PlantSimEngine.ModelList

PlantSimEngine.Status

Base.copy

Base.copy

Base.copy

Base.getindex

PlantSimEngine.EF

PlantSimEngine.NRMSE

PlantSimEngine.RMSE

PlantSimEngine.add_model_vars

PlantSimEngine.check_dimensions

PlantSimEngine.convert_vars

PlantSimEngine.dep

PlantSimEngine.diff_vars

PlantSimEngine.dr

PlantSimEngine.draw_guide

PlantSimEngine.draw_panel

PlantSimEngine.drop_process

PlantSimEngine.flatten_vars

PlantSimEngine.get_attr_i

PlantSimEngine.homogeneous_ts_kwargs

PlantSimEngine.homogeneous_ts_kwargs

PlantSimEngine.init_mtg_models!

PlantSimEngine.init_status!

PlantSimEngine.init_variables

PlantSimEngine.init_variables_manual

PlantSimEngine.inputs

PlantSimEngine.is_initialized

PlantSimEngine.outputs

PlantSimEngine.pre_allocate_attr!

PlantSimEngine.pull_status!

PlantSimEngine.pull_status_one_step!

PlantSimEngine.run!

PlantSimEngine.search_inputs_in_output

PlantSimEngine.soft_dependencies

PlantSimEngine.status

PlantSimEngine.to_initialize

PlantSimEngine.to_initialize

PlantSimEngine.traverse_dependency_tree

PlantSimEngine.traverse_dependency_tree!

PlantSimEngine.traverse_dependency_tree!

PlantSimEngine.update_mtg_models!

PlantSimEngine.variables

PlantSimEngine.variables

PlantSimEngine.variables

PlantSimEngine.variables_typed

PlantSimEngine.vars_not_init_

Tables.schema

PlantSimEngine.@process

API documentation

DataFrames.DataFrame — Method

DataFrame(components <: AbstractArray{<:ModelList})
DataFrame(components <: AbstractDict{N,<:ModelList})

Fetch the data from a ModelList (or an Array/Dict of) status into a DataFrame.

Examples

using PlantSimEngine
using DataFrames

# Creating a ModelList
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

# Converting to a DataFrame
df = DataFrame(models)

# Converting to a Dict of ModelLists
models = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    ),
    "InterNode" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    )
)

# Converting to a DataFrame
df = DataFrame(models)

DataFrames.DataFrame — Method

DataFrame(components::ModelList{T,S}) where {T,S<:AbstractDict}

Implementation of DataFrame for a ModelList model with one time step.

DataFrames.DataFrame — Method

DataFrame(components::ModelList{T,<:TimeStepTable})

Implementation of DataFrame for a ModelList model with several time steps.

PlantMeteo.TimeStepTable — Method

TimeStepTable{Status}(df::DataFrame)

Method to build a TimeStepTable (from PlantMeteo.jl) from a DataFrame, but with each row being a Status.

Note

ModelList uses TimeStepTable{Status} by default (see examples below).

Examples

using PlantSimEngine, DataFrames

# A TimeStepTable from a DataFrame:
df = DataFrame(
    Tₗ=[25.0, 26.0],
    PPFD=[1000.0, 1200.0],
    Cₛ=[400.0, 400.0],
    Dₗ=[1.0, 1.2],
)
TimeStepTable{Status}(df)

# A leaf with several values for at least one of its variable will automatically use 
# TimeStepTable{Status} with the time steps:
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

# The status of the leaf is a TimeStepTable:
status(models)

# Of course we can also create a TimeStepTable with Status manually:
TimeStepTable(
    [
        Status(Tₗ=25.0, PPFD=1000.0, Cₛ=400.0, Dₗ=1.0),
        Status(Tₗ=26.0, PPFD=1200.0, Cₛ=400.0, Dₗ=1.2),
    ]
)

PlantSimEngine.AbstractModel — Type

Abstract model type. All models are subtypes of this one.

PlantSimEngine.DependencyTree — Type

DependencyTree{T}(roots::T, not_found::Dict{Symbol,DataType})

A tree of dependencies between models.

Arguments

roots::T: the root nodes of the tree.

not_found::Dict{Symbol,DataType}: the models that were not found in the tree.

PlantSimEngine.ModelList — Type

ModelList(models::M, status::S)
ModelList(;
    status=nothing,
    init_fun::Function=init_fun_default,
    type_promotion=nothing,
    variables_check=true,
    kwargs...
)

List the models for a simulation (models), and does all boilerplate for variable initialization, type promotion, time steps handling.

Note

The status field depends on the input models. You can get the variables needed by a model using variables on the instantiation of a model. You can also use inputs and outputs instead.

Arguments

models: a list of models. Usually given as a NamedTuple, but can be any other structure that

implements getproperty.

status: a structure containing the initializations for the variables of the models. Usually a NamedTuple

when given as a kwarg, or any structure that implements the Tables interface from Tables.jl (e.g. DataFrame).

init_fun: a function that initializes the status based on a vector of NamedTuples (see details).

type_promotion: optional type conversion for the variables with default values.

nothing by default, i.e. no conversion. Note that conversion is not applied to the variables input by the user as kwargs (need to do it manually). Should be provided as a Dict with current type as keys and new type as values.

variables_check=true: check that all needed variables are initialized by the user.

kwargs: the models, named after the process they simulate.

Details

The argument init_fun is set by default to init_fun_default which initializes the status with a TimeStepTable of Status structures.

If you change init_fun by another function, make sure the type you are using (i.e. in place of TimeStepTable) implements the Tables.jl interface (e.g. DataFrame does). And if you still use TimeStepTable but only change Status, make sure the type you give is indexable using the dot synthax (e.g. x.var).

If you need to input a custom Type for the status and make your users able to only partially initialize the status field in the input, you'll have to implement a method for add_model_vars!, a function that adds the models variables to the type in case it is not fully initialized.

Examples

We'll use the dummy models from the dummy.jl in the examples folder of the package. It implements three dummy processes: Process1Model, Process2Model and Process3Model, with one model implementation each: Process1Model, Process2Model and Process3Model.

julia> using PlantSimEngine;

Including an example script that implements dummy processes and models:

julia> include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model());
[ Info: Some variables must be initialized before simulation: (process1 = (:var1, :var2), process2 = (:var1,)) (see `to_initialize()`)

julia> typeof(models)
ModelList{NamedTuple{(:process1, :process2, :process3), Tuple{Process1Model, Process2Model, Process3Model}}, TimeStepTable{Status{(:var4, :var5, :var6, :var1, :var3, :var2), NTuple{6, Base.RefValue{Float64}}}}}

No variables were given as keyword arguments, that means that the status of the ModelList is not set yet, and all variables are initialized to their default values given in the inputs and outputs (usually typemin(Type), i.e. -Inf for floating point numbers). This component cannot be simulated yet.

To know which variables we need to initialize for a simulation, we use to_initialize:

julia> to_initialize(models)
(process1 = (:var1, :var2), process2 = (:var1,))

We can now provide values for these variables in the status field, and simulate the ModelList, e.g. for process3 (coupled with process1 and process2):

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=(var1=15.0, var2=0.3));

julia> meteo = Atmosphere(T = 22.0, Wind = 0.8333, P = 101.325, Rh = 0.4490995);

julia> run!(models,meteo)

julia> models[:var6]
1-element Vector{Float64}:
 58.0138985

If we want to use special types for the variables, we can use the type_promotion argument:

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=(var1=15.0, var2=0.3), type_promotion = Dict(Float64 => Float32));

We used type_promotion to force the status into Float32:

julia> [typeof(models[i][1]) for i in keys(status(models))]
6-element Vector{DataType}:
 Float32
 Float32
 Float32
 Float64
 Float64
 Float32

But we see that only the default variables (the ones that are not given in the status arguments) were converted to Float32, the two other variables that we gave were not converted. This is because we want to give the ability to users to give any type for the variables they provide in the status. If we want all variables to be converted to Float32, we can pass them as Float32:

julia> models = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=(var1=15.0f0, var2=0.3f0), type_promotion = Dict(Float64 => Float32));

We used type_promotion to force the status into Float32:

julia> [typeof(models[i][1]) for i in keys(status(models))]
6-element Vector{DataType}:
 Float32
 Float32
 Float32
 Float32
 Float32
 Float32

We can also use DataFrame as the status type:

julia> using DataFrames;

julia> df = DataFrame(:var1 => [13.747, 13.8], :var2 => [1.0, 1.0]);

julia> m = ModelList(process1=Process1Model(1.0), process2=Process2Model(), process3=Process3Model(), status=df, init_fun=x -> DataFrame(x));

Note that we use init_fun to force the status into a DataFrame, otherwise it would be automatically converted into a TimeStepTable{Status}.

julia> status(m)
2×6 DataFrame
 Row │ var4     var5     var6     var1     var3     var2    
     │ Float64  Float64  Float64  Float64  Float64  Float64 
─────┼──────────────────────────────────────────────────────
   1 │    -Inf     -Inf     -Inf   13.747     -Inf      1.0
   2 │    -Inf     -Inf     -Inf   13.8       -Inf      1.0

Note that computations will be slower using DataFrame, so if performance is an issue, use TimeStepTable instead (or a NamedTuple as shown in the example).

PlantSimEngine.Status — Type

Status(vars)

Status type used to store the values of the variables during simulation. It is mainly used as the structure to store the variables in the TimeStepRow of a TimeStepTable (see PlantMeteo.jl docs) of a ModelList.

Most of the code is taken from MasonProtter/MutableNamedTuples.jl, so Status is a MutableNamedTuples with a few modifications, so in essence, it is a stuct that stores a NamedTuple of the references to the values of the variables, which makes it mutable.

Examples

A leaf with one value for all variables will make a status with one time step:

julia> st = PlantSimEngine.Status(Rₛ=13.747, sky_fraction=1.0, d=0.03, PPFD=1500.0);

All these indexing methods are valid:

julia> st[:Rₛ]
13.747

julia> st.Rₛ
13.747

julia> st[1]
13.747

Setting a Status variable is very easy:

julia> st[:Rₛ] = 20.0
20.0

julia> st.Rₛ = 21.0
21.0

julia> st[1] = 22.0
22.0

Base.copy — Method

Base.copy(l::AbstractArray{<:ModelList})

Copy an array-alike of ModelList

Base.copy — Method

Base.copy(l::AbstractDict{N,<:ModelList} where N)

Copy a Dict-alike ModelList

Base.copy — Method

Base.copy(l::ModelList)
Base.copy(l::ModelList, status)

Copy a ModelList, eventually with new values for the status.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

# Create a model list:
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

# Copy the model list:
ml2 = copy(models)

# Copy the model list with new status:
ml3 = copy(models, TimeStepTable([Status(var1=20.0, var2=0.5))])

Base.getindex — Method

getindex(component<:ModelList, key::Symbol)
getindex(component<:ModelList, key)

Indexing a component models structure: - with an integer, will return the status at the ith time-step - with anything else (Symbol, String) will return the required variable from the status

Examples

using PlantSimEngine

lm = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status = (var1=[15.0, 16.0], var2=0.3)
);

lm[:var1] # Returns the value of the Tₗ variable
lm[2]  # Returns the status at the second time-step
lm[2][:var1] # Returns the value of Tₗ at the second time-step
lm[:var1][2] # Equivalent of the above

# output
16.0

PlantSimEngine.EF — Method

EF(obs,sim)

Returns the Efficiency Factor between observations obs and simulations sim using NSE (Nash-Sutcliffe efficiency) model. More information can be found at https://en.wikipedia.org/wiki/Nash%E2%80%93Sutcliffemodelefficiency_coefficient.

The closer to 1 the better.

Examples

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

EF(obs, sim)

PlantSimEngine.NRMSE — Method

NRMSE(obs,sim)

Returns the Normalized Root Mean Squared Error between observations obs and simulations sim. Normalization is performed using division by observations range (max-min).

Examples

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

NRMSE(obs, sim)

PlantSimEngine.RMSE — Method

RMSE(obs,sim)

Returns the Root Mean Squared Error between observations obs and simulations sim.

The closer to 0 the better.

Examples

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

RMSE(obs, sim)

PlantSimEngine.add_model_vars — Method

add_model_vars(x)

Check which variables in x are not initialized considering a set of models and the variables needed for their simulation. If some variables are uninitialized, initialize them to their default values.

This function needs to be implemented for each type of x (please do it if you need it).

Careful, the function mutates x in place for performance. We don't put the ! in the name just because it also returns it (impossible to mutate when x is nothing)

PlantSimEngine.check_dimensions — Method

check_dimensions(component,weather)
check_dimensions(status,weather)

Checks if a component status (or a status directly) and the weather have the same length, or if they can be recycled (length 1 for one of them).

Examples

using PlantSimEngine, PlantMeteo

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

# Creating a dummy weather:
w = Atmosphere(T = 20.0, Rh = 0.5, Wind = 1.0)

# Creating a dummy component:
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=[15.0, 16.0], var2=0.3)
)

# Checking that the number of time-steps are compatible (here, they are, it returns nothing):
PlantSimEngine.check_dimensions(models, w) 

# Creating a dummy weather with 3 time-steps:
w = Weather([
    Atmosphere(T = 20.0, Rh = 0.5, Wind = 1.0),
    Atmosphere(T = 25.0, Rh = 0.5, Wind = 1.0),
    Atmosphere(T = 30.0, Rh = 0.5, Wind = 1.0)
])

# Checking that the number of time-steps are compatible (here, they are not, it throws an error):
PlantSimEngine.check_dimensions(models, w)

# output
ERROR: DimensionMismatch: Component status should have the same number of time-steps (2) than weather data (3).

PlantSimEngine.convert_vars — Method

convert_vars(type_promotion::Dict{DataType,DataType}, ref_vars)
convert_vars(type_promotion::Nothing, ref_vars)

Convert the status variables to the type specified in the type promotion dictionary.

Examples

If we want all the variables that are Reals to be Float32, we can use:

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

ref_vars = init_variables(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
)
type_promotion = Dict(Real => Float32)

PlantSimEngine.convert_vars(type_promotion, ref_vars.process3)

PlantSimEngine.dep — Method

dep(models::ModelList; verbose::Bool=true)

Get the model dependency tree given a ModelList. If one tree is returned, then all models are coupled. If several trees are returned, then only the models inside each tree are coupled, and the models in different trees are not coupled.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status=(var1=15.0, var2=0.3)
)

dep(models)

# or directly with the processes:
vars = (
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    process4=Process4Model(),
    process5=Process5Model(),
    process6=Process6Model(),
    process7=Process7Model(),
)

dep(;vars...)

PlantSimEngine.diff_vars — Method

diff_vars(x, y)

Returns the names of variables that have different values in x and y.

PlantSimEngine.dr — Method

dr(obs,sim)

Returns the Willmott’s refined index of agreement dᵣ. Willmot et al. 2011. A refined index of model performance. https://rmets.onlinelibrary.wiley.com/doi/10.1002/joc.2419

The closer to 1 the better.

Examples

using PlantSimEngine

obs = [1.0, 2.0, 3.0]
sim = [1.1, 2.1, 3.1]

dr(obs, sim)

PlantSimEngine.draw_guide — Method

draw_guide(h, w, prefix, isleaf, guides)

Draw the line guide for one node of the dependency tree.

PlantSimEngine.draw_panel — Method

draw_panel(node, tree, prefix, dep_tree_guides, parent; title="Soft-coupled model")

Draw the panels for all dependencies

PlantSimEngine.drop_process — Method

drop_process(proc_vars, process)

Return a new NamedTuple with the process process removed from the NamedTuple proc_vars.

Arguments

proc_vars::NamedTuple: the NamedTuple from which we want to remove the process process.

process::Symbol: the process we want to remove from the NamedTuple proc_vars.

Returns

A new NamedTuple with the process process removed from the NamedTuple proc_vars.

Example

julia> drop_process((a = 1, b = 2, c = 3), :b)
(a = 1, c = 3)

julia> drop_process((a = 1, b = 2, c = 3), (:a, :c))
(b = 2,)

PlantSimEngine.flatten_vars — Method

flatten_vars(vars)

Return a set of the variables in the vars dictionary.

Arguments

vars::Dict{Symbol, Tuple{Symbol, Vararg{Symbol}}}: a dict of process => symbols of inputs.

Returns

A set of the variables in the vars dictionary.

Example

julia> flatten_vars(Dict(:process1 => (:var1, :var2), :process2 => (:var3, :var4)))
Set{Symbol} with 4 elements:
  :var4
  :var3
  :var2
  :var1

PlantSimEngine.get_attr_i — Method

get_attr_i(node, attr, i<:Nothing)
get_attr_i(node, attr, i)

Get node attribute at ith value or only value if i<:Nothing

PlantSimEngine.homogeneous_ts_kwargs — Method

homogeneous_ts_kwargs(kwargs)

By default, the function returns its argument.

PlantSimEngine.homogeneous_ts_kwargs — Method

kwargs_to_timestep(kwargs::NamedTuple{N,T}) where {N,T}

Takes a NamedTuple with optionnaly vector of values for each variable, and makes a vector of NamedTuple, with each being a time step. It is used to be able to e.g. give constant values for all time-steps for one variable.

Examples

PlantSimEngine.homogeneous_ts_kwargs((Tₗ=[25.0, 26.0], PPFD=1000.0))

PlantSimEngine.init_mtg_models! — Function

init_mtg_models!(
    mtg::MultiScaleTreeGraph.Node,
    models::Dict{String,<:ModelList},
    i=nothing;
    verbose=true,
    attr_name=:models
)

initialize the components of an MTG (i.e. nodes) with the corresponding models.

The function checks if the models associated to each component of the MTG are fully initialized, and if not, tries to initialize the variables using the MTG attributes with the exact same name, and if not found, returns an error.

Arguments

mtg::MultiScaleTreeGraph.Node: the MTG tree.

models::Dict{String,ModelList}: a dictionary of models named by components names

i=nothing: the time-step to initialize. If nothing, initialize all the time-steps.

verbose = true: return information during the processes

attr_name = :models: the node attribute name used to store the models

Examples

using PlantSimEngine, MultiScaleTreeGraph

# Make an MTG:
mtg = Node(MultiScaleTreeGraph.NodeMTG("/", "Plant", 1, 1))
internode = Node(mtg, MultiScaleTreeGraph.NodeMTG("/", "Internode", 1, 2))
leaf = Node(mtg, MultiScaleTreeGraph.NodeMTG("<", "Leaf", 1, 2))
leaf[:var1] = [15.0, 16.0]
leaf[:var2] = 0.3

# Declare our models:
models = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    )
)

# Checking which variables are needed for our models:
[component => to_initialize(model) for (component, model) in models]
# OK we need to initialize Rₛ, sky_fraction and the PPFD

# We could compute them directly inside the MTG from available variables instead of 
# giving them as initialisations:
transform!(
    mtg,
    :var1 => (x -> x .+ 2.0) => :var2,
    ignore_nothing = true
)

# Initialising all components with their corresponding models and initialisations at time-step 1:
init_mtg_models!(mtg, models, 1)
# Note that this is possible only because the initialisation values are found in the MTG.
# If the initialisations are constant values between components, we can directly initialize
# them in the models definition (as we do in the begining).

PlantSimEngine.init_status! — Method

init_status!(object::Dict{String,ModelList};vars...)
init_status!(component::ModelList;vars...)

Initialise model variables for components with user input.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

models = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    ),
    "InterNode" => ModelList(
        process1=Process1Model(1.0),
    )
)

init_status!(models, var1=1.0 , var2=2.0)
status(models["Leaf"])

PlantSimEngine.init_variables — Method

init_variables(models...)

Initialized model variables with their default values. The variables are taken from the inputs and outputs of the models.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

init_variables(Process1Model(2.0))
init_variables(process1=Process1Model(2.0), process2=Process2Model())

PlantSimEngine.init_variables_manual — Method

init_variables_manual(models...;vars...)

Return an initialisation of the model variables with given values.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model()
)

PlantSimEngine.init_variables_manual(status(models), (var1=20.0,))

PlantSimEngine.inputs — Method

inputs(model::AbstractModel)
inputs(...)

Get the inputs of one or several models.

Returns an empty tuple by default for AbstractModels (no inputs) or Missing models.

Examples

using PlantSimEngine;

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

inputs(Process1Model(1.0))

# output
(:var1, :var2)

PlantSimEngine.is_initialized — Method

is_initialized(m::T) where T <: ModelList
is_initialized(m::T, models...) where T <: ModelList

Check if the variables that must be initialized are, and return true if so, and false and an information message if not.

Note

There is no way to know before-hand which process will be simulated by the user, so if you have a component with a model for each process, the variables to initialize are always the smallest subset of all, meaning it is considered the user will simulate the variables needed for other models.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model()
)

is_initialized(models)

PlantSimEngine.outputs — Method

outputs(model::AbstractModel)
outputs(...)

Get the outputs of one or several models.

Returns an empty tuple by default for AbstractModels (no outputs) or Missing models.

Examples

using PlantSimEngine;

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

outputs(Process1Model(1.0))

# output
(:var3,)

PlantSimEngine.pre_allocate_attr! — Method

pre_allocate_attr!(node, nsteps; attr_name = :models)

Pre-allocate the node attributes based on the status of a component model and a given number of simulation steps.

PlantSimEngine.pull_status! — Method

pull_status!(node)

Copy the status of a node's component models (e.g. the outputs of a [ModelList]@ref simulation) into the MTG attributes. This function is used when we need to compute further the simulation outputs with e.g. transform! (see MultiScaleTreeGraph.jl docs).

Notes

Carefull, this function makes a copy, so the values are then present at two locations (can take a lot of memory space if using several plants).

Examples

using PlantSimEngine, MultiScaleTreeGraph

# Make an MTG:
mtg = Node(MultiScaleTreeGraph.NodeMTG("/", "Plant", 1, 1))
internode = Node(mtg, MultiScaleTreeGraph.NodeMTG("/", "Internode", 1, 2))
leaf = Node(mtg, MultiScaleTreeGraph.NodeMTG("<", "Leaf", 1, 2))
leaf[:var1] = [15.0, 16.0]
leaf[:var2] = 0.3

# Declare our models:
models = Dict(
    "Leaf" => ModelList(
        process1=Process1Model(1.0),
        process2=Process2Model(),
        process3=Process3Model()
    )
)

# Initialising all components with their corresponding models and initialisations:
init_mtg_models!(mtg, models, 1)

# Make a simulation
transform!(mtg, :models => (x -> run!(x, meteo)), ignore_nothing = true)
# Pull the simulation results into the MTG attributes:
transform!(mtg, PlantSimEngine.pull_status!)
# Now the simulated variables are available from the MTG attributes field:
names(mtg)

PlantSimEngine.pull_status_one_step! — Method

pull_status_one_step!(node, step; attr_name = :models)

Copy the status of a node's ModelList (i.e. the outputs of the simulations) into the pre-allocated MTG attributes, i.e. one value per step.

See pre_allocate_attr! for the pre-allocation step.

PlantSimEngine.run! — Function

run!(object, meteo, constants, extra=nothing; check=true)

Run the simulation for each model in the model list in the correct order, i.e. respecting the dependency tree.

If several time-steps are given, the models are run sequentially for each time-step.

Arguments

object: a ModelList, an array or dict of ModelList, or an MTG.

meteo: a PlantMeteo.TimeStepTable of

PlantMeteo.Atmosphere or a single PlantMeteo.Atmosphere.

constants: a PlantMeteo.Constants object, or a NamedTuple of constant keys and values.

extra: extra parameters.

Returns

Modifies the status of the object in-place. Users may retrieve the results from the object using the status function (see examples).

Details

The models are run according to the dependency tree. If a model has a soft dependency on another model (i.e. its inputs are computed by another model), the other model is run first. If a model has several soft dependencies, the parents (the soft dependencies) are always computed first.

Example

Import the packages:

julia> using PlantSimEngine, PlantMeteo;

Load the dummy models given as example in the package:

julia> include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

Create a model list:

julia> models = ModelList(Process1Model(1.0), Process2Model(), Process3Model(), status = (var1=1.0, var2=2.0));

Create meteo data:

julia> meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_f=300.0);

Run the simulation:

julia> run!(models, meteo);

Get the results:

julia> (models[:var4],models[:var6])
([12.0], [41.95])

PlantSimEngine.search_inputs_in_output — Method

search_inputs_in_output(process, inputs, outputs)

Return a dictionary with the soft dependencies of the processes in the dependency tree d. A soft dependency is a dependency that is not explicitely defined in the model, but that can be inferred from the inputs and outputs of the processes.

Arguments

process::Symbol: the process for which we want to find the soft dependencies.

inputs::Dict{Symbol, Vector{Pair{Symbol}, Tuple{Symbol, Vararg{Symbol}}}}: a dict of process => symbols of inputs per process.

outputs::Dict{Symbol, Tuple{Symbol, Vararg{Symbol}}}: a dict of process => symbols of outputs per process.

Details

The inputs (and similarly, outputs) give the inputs of each process, classified by the process it comes from. It can come from itself (its own inputs), or from another process that is a hard-dependency.

Returns

A dictionary with the soft dependencies for the processes.

Example

in_ = Dict(
    :process3 => [:process3=>(:var4, :var5), :process2=>(:var1, :var3), :process1=>(:var1, :var2)],
    :process4 => [:process4=>(:var0,)],
    :process6 => [:process6=>(:var7, :var9)],
    :process5 => [:process5=>(:var5, :var6)],
)

out_ = Dict(
    :process3 => Pair{Symbol}[:process3=>(:var4, :var6), :process2=>(:var4, :var5), :process1=>(:var3,)],
    :process4 => [:process4=>(:var1, :var2)],
    :process6 => [:process6=>(:var8,)],
    :process5 => [:process5=>(:var7,)],
)

search_inputs_in_output(:process3, in_, out_)
(process4 = (:var1, :var2),)

PlantSimEngine.soft_dependencies — Method

soft_dependencies(d::DependencyTree)

Return a DependencyTree with the soft dependencies of the processes in the dependency tree d. A soft dependency is a dependency that is not explicitely defined in the model, but that can be inferred from the inputs and outputs of the processes.

Arguments

d::DependencyTree: the hard-dependency tree.

Example

using PlantSimEngine

# Load the dummy models given as example in the package:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

# Create a model list:
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    process4=Process4Model(),
    process5=Process5Model(),
    process6=Process6Model(),
)

# Create the hard-dependency tree:
hard_dep = hard_dependencies(models.models, verbose=true)

# Get the soft dependencies tree:
soft_dep = soft_dependencies(hard_dep)

PlantSimEngine.status — Method

status(m)
status(m::AbstractArray{<:ModelList})
status(m::AbstractDict{T,<:ModelList})

Get a ModelList status, i.e. the state of the input (and output) variables.

See also is_initialized and to_initialize

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

# Create a ModelList
models = ModelList(
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    status = (var1=[15.0, 16.0], var2=0.3)
);

status(models)

# Or just one variable:
status(models,:var1)

# Or the status at the ith time-step:
status(models, 2)

# Or even more simply:
models[:var1]
# output
2-element Vector{Float64}:
 15.0
 16.0

PlantSimEngine.to_initialize — Method

to_initialize(v::T, vars...) where T <: Union{Missing,AbstractModel}
to_initialize(m::T)  where T <: ModelList
to_initialize(m::DependencyTree)

Return the variables that must be initialized providing a set of models and processes. The function takes into account model coupling and only returns the variables that are needed considering that some variables that are outputs of some models are used as inputs of others.

Examples

using PlantSimEngine

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

to_initialize(process1=Process1Model(1.0), process2=Process2Model())

# Or using a component directly:
models = ModelList(process1=Process1Model(1.0), process2=Process2Model())
to_initialize(models)

m = ModelList(
    (
        process1=Process1Model(1.0),
        process2=Process2Model()
    ),
    Status(var1 = 5.0, var2 = -Inf, var3 = -Inf, var4 = -Inf, var5 = -Inf)
)

PlantSimEngine.to_initialize — Method

to_initialize(m::AbstractDependencyNode)

Return the variables that must be initialized providing a set of models and processes. The function just returns the inputs and outputs of each model, with their default values. To take into account model coupling, use the function at an upper-level instead, i.e. to_initialize(m::ModelList) or to_initialize(m::DependencyTree).

PlantSimEngine.traverse_dependency_tree! — Method

traverse_dependency_tree(node::HardDependencyNode, f::Function, var::Vector)

Apply function f to node, and then its children (hard-dependency nodes).

Mutate the vector var by pushing a pair of the node process name and the result of the function f.

PlantSimEngine.traverse_dependency_tree! — Method

traverse_dependency_tree(node::SoftDependencyNode, f::Function, var::Vector; visit_hard_dep=true)

Apply function f to node, visit its hard dependency nodes (if visit_hard_dep=true), and then its soft dependency children.

Mutate the vector var by pushing a pair of the node process name and the result of the function f.

PlantSimEngine.traverse_dependency_tree — Method

traverse_dependency_tree(tree::DependencyTree, f::Function, visit_hard_dep=true)

Traverse the dependency tree and apply the function f to each node. The first-level soft-dependencies are traversed first, then their hard-dependencies (if visit_hard_dep=true), and then the children of the soft-dependencies.

Return a vector of pairs of the node and the result of the function f.

Example

using PlantSimEngine

function f(node)
    node.value
end

include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))

vars = (
    process1=Process1Model(1.0),
    process2=Process2Model(),
    process3=Process3Model(),
    process4=Process4Model(),
    process5=Process5Model(),
    process6=Process6Model(),
    process7=Process7Model(),
)

tree = dep(vars)

traverse_dependency_tree(tree, f)

PlantSimEngine.update_mtg_models! — Method

update_mtg_models!(mtg::MultiScaleTreeGraph.Node, i, attr_name::Symbol)

Update the mtg models initialisations by using the ith time-step. The mtg is considered fully initialized already once, so init_mtg_models! must be called before update_mtg_models!.

The values are updated only for node attributes in to_init. Those attributes must have several time-steps, i.e. indexable by 1:n time-steps.

PlantSimEngine.variables — Method

variables(pkg::Module)

Returns a dataframe of all variables, their description and units in a package that has PlantSimEngine as a dependency (if implemented by the authors).

Note to developers

Developers of a package that depends on PlantSimEngine should put a csv file in "data/variables.csv", then this file will be returned by the function.

Examples

Here is an example with the PlantBiophysics package:

#] add PlantBiophysics
using PlantBiophysics
variables(PlantBiophysics)

PlantSimEngine.variables — Method

variables(m::AbstractDependencyNode)

Returns a tuple with the name of the inputs and outputs variables needed by a model in a dependency tree.

PlantSimEngine.variables — Method

variables(model)
variables(model, models...)

Returns a tuple with the name of the variables needed by a model, or a union of those variables for several models.

Note

Each model can (and should) have a method for this function.

using PlantSimEngine;

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

variables(Process1Model(1.0))

variables(Process1Model(1.0), Process2Model())

# output

5-element Vector{Symbol}:
 :var1
 :var2
 :var3
 :var4
 :var5

See also

inputs, outputs and variables_typed

PlantSimEngine.variables_typed — Method

variables_typed(model)
variables_typed(model, models...)

Returns a named tuple with the name and the types of the variables needed by a model, or a union of those for several models.

Examples

using PlantSimEngine;

# Including an example script that implements dummy processes and models:
include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"));

PlantSimEngine.variables_typed(Process1Model(1.0))
(var1 = Float64, var2 = Float64, var3 = Float64)

PlantSimEngine.variables_typed(Process1Model(1.0), Process2Model())

# output
(var4 = Float64, var5 = Float64, var1 = Float64, var2 = Float64, var3 = Float64)

See also

inputs, outputs and variables

PlantSimEngine.vars_not_init_ — Method

vars_not_init_(st<:Status, var_names)

Get which variable is not properly initialized in the status struct.

Tables.schema — Method

Tables.schema(m::TimeStepTable{Status})

Create a schema for a TimeStepTable{Status}.

PlantSimEngine.@process — Macro

@process(process::String, doc::String=""; verbose::Bool=true)

This macro generate the abstract type and some boilerplate code for the simulation of a process, along with its documentation. It also prints out a short tutorial for implementing a model if verbose=true.

The abstract process type is then used as a supertype of all models implementations for the process, and is named "Abstract<ProcessName>Model", e.g. AbstractGrowthModel for a process called growth.

The first argument to @process is the new process name, the second is any additional documentation that should be added to the Abstract<ProcessName>Model type, and the third determines whether the short tutorial should be printed or not.

Examples

@process "dummy_process" "This is a dummy process that shall not be used"